Specification for the observation error model (OEM).

The FLoem class stores the method, arguments and observations that define the way observations are collected from an operating model at each time step in the management procedure. This class extends mseCtrl through the addition of two lists used to gather past and new observations, and deviances to use on each step in the observation process.

Usage

FLoem(...)

FLoem(...)

observations(object, ...)

# S4 method for class 'FLoem'
observations(object, ...)

observations(object, i) <- value

# S4 method for class 'FLoem,missing,list'
observations(object) <- value

# S4 method for class 'FLoem,ANY,FLStock'
observations(object, i) <- value

# S4 method for class 'FLoem'
deviances(object, ...)

# S4 method for class 'FLoem,list'
deviances(object, ...) <- value

# S4 method for class 'FLoem'
show(object)

# S4 method for class 'FLoem'
iter(obj, iter)

# S4 method for class 'FLoem,FLoem'
combine(x, y, ..., check = FALSE)

Arguments

...

additional argument list that might never be used

object

object of relevant class (see signature of method)

value

the new object

Slots

method

The function to be run in the module call, class function.

args

Arguments to be passed to the method, of class list.

observations

Past observations, class list.

deviances

Observation deviances, class list.

Accessors

All slots in the class have accessor and replacement methods defined that allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot. A numeric vector can also be used when replacing FLQuant slots, and the vector will be used to substitute the values in the slot, but not its other attributes.

Constructor

A construction method exists for this class that can take named arguments for any of its slots. All slots are then created to match the requirements of the class validity. If an unnamed FLQuant object is provided, this is used for sizing, but not for populating any slot.

Examples

data(plesim)
deviances(oem)
#> $stk
#> $ catch.n 
#> An object of class "FLQuant"
#> iters:  100 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1960         1961         1962         1963         1964        
#>   1  0.997(0.176) 0.976(0.191) 1.098(0.189) 0.984(0.220) 1.015(0.190)
#>   2  1.002(0.202) 1.013(0.220) 1.008(0.193) 0.964(0.196) 0.988(0.220)
#>   3  1.001(0.189) 0.947(0.175) 0.961(0.169) 0.972(0.254) 1.036(0.210)
#>   4  1.019(0.184) 0.999(0.196) 1.045(0.275) 1.014(0.189) 1.022(0.168)
#>   5  1.048(0.194) 0.992(0.210) 0.987(0.206) 1.035(0.198) 1.000(0.213)
#>   6  0.980(0.179) 0.989(0.174) 0.957(0.196) 1.026(0.215) 0.979(0.168)
#>   7  0.995(0.219) 0.984(0.186) 1.026(0.211) 0.979(0.201) 1.014(0.201)
#>   8  1.036(0.186) 1.019(0.211) 1.019(0.217) 1.030(0.208) 0.956(0.153)
#>   9  0.991(0.220) 0.992(0.173) 1.012(0.222) 0.998(0.249) 0.968(0.211)
#>   10 1.015(0.185) 0.989(0.210) 1.005(0.165) 0.990(0.222) 0.958(0.207)
#>     year
#> age  1965         1966         1967         1968         1969        
#>   1  0.962(0.206) 1.026(0.205) 1.020(0.228) 0.994(0.223) 0.987(0.169)
#>   2  0.994(0.186) 1.038(0.174) 0.960(0.216) 0.971(0.171) 1.023(0.258)
#>   3  1.020(0.196) 0.977(0.214) 0.993(0.200) 0.985(0.190) 1.007(0.208)
#>   4  1.018(0.183) 1.029(0.208) 0.979(0.207) 1.013(0.184) 0.990(0.194)
#>   5  0.997(0.168) 0.973(0.198) 1.004(0.186) 0.995(0.208) 0.993(0.185)
#>   6  0.981(0.200) 0.956(0.172) 0.975(0.174) 0.970(0.190) 1.021(0.230)
#>   7  1.001(0.159) 0.966(0.190) 1.006(0.211) 0.974(0.183) 1.017(0.177)
#>   8  0.976(0.231) 0.970(0.135) 1.035(0.186) 1.004(0.201) 1.013(0.202)
#>   9  1.031(0.276) 1.001(0.183) 1.025(0.187) 0.971(0.203) 1.035(0.184)
#>   10 1.013(0.180) 0.993(0.156) 1.017(0.198) 0.953(0.208) 0.996(0.189)
#>     year
#> age  1970         1971         1972         1973         1974        
#>   1  1.002(0.174) 0.965(0.177) 0.996(0.153) 1.013(0.178) 1.012(0.202)
#>   2  0.998(0.173) 0.966(0.182) 0.984(0.182) 1.008(0.171) 0.983(0.209)
#>   3  0.993(0.239) 0.985(0.206) 0.995(0.213) 1.013(0.209) 1.008(0.191)
#>   4  0.979(0.232) 0.998(0.185) 1.026(0.211) 0.966(0.217) 0.970(0.205)
#>   5  1.008(0.182) 0.997(0.236) 1.009(0.144) 1.021(0.184) 0.982(0.188)
#>   6  0.976(0.191) 1.004(0.181) 0.963(0.196) 0.961(0.187) 0.968(0.183)
#>   7  1.004(0.192) 1.019(0.208) 0.954(0.175) 0.992(0.197) 1.010(0.231)
#>   8  0.948(0.187) 0.960(0.224) 0.992(0.152) 0.988(0.221) 0.989(0.177)
#>   9  1.030(0.203) 0.975(0.162) 0.997(0.186) 0.984(0.184) 1.015(0.222)
#>   10 1.038(0.182) 0.995(0.185) 0.982(0.210) 1.003(0.220) 0.976(0.186)
#>     year
#> age  1975         1976         1977         1978         1979        
#>   1  0.958(0.216) 0.999(0.192) 0.971(0.230) 1.030(0.184) 0.984(0.203)
#>   2  0.991(0.213) 1.023(0.226) 0.996(0.235) 0.985(0.197) 1.018(0.176)
#>   3  1.029(0.207) 1.015(0.197) 1.017(0.189) 1.004(0.224) 0.998(0.195)
#>   4  1.006(0.205) 1.004(0.175) 1.013(0.222) 0.983(0.193) 0.985(0.210)
#>   5  1.056(0.218) 0.971(0.175) 0.973(0.244) 1.018(0.205) 1.042(0.188)
#>   6  0.992(0.200) 0.978(0.202) 1.050(0.219) 1.039(0.193) 0.970(0.201)
#>   7  0.967(0.172) 0.983(0.197) 0.989(0.203) 0.969(0.197) 1.030(0.200)
#>   8  0.997(0.160) 0.990(0.183) 0.970(0.191) 0.984(0.203) 1.013(0.242)
#>   9  0.981(0.193) 0.976(0.182) 0.981(0.208) 1.042(0.219) 1.020(0.176)
#>   10 0.999(0.181) 0.976(0.194) 0.965(0.185) 0.963(0.218) 1.003(0.189)
#>     year
#> age  1980         1981         1982         1983         1984        
#>   1  1.082(0.209) 0.994(0.217) 1.076(0.224) 0.997(0.217) 0.985(0.185)
#>   2  1.007(0.183) 0.997(0.200) 0.991(0.147) 1.041(0.240) 0.956(0.200)
#>   3  1.011(0.212) 1.021(0.162) 1.001(0.172) 1.001(0.200) 1.025(0.160)
#>   4  0.993(0.218) 0.987(0.168) 1.000(0.185) 0.969(0.221) 1.024(0.232)
#>   5  0.990(0.159) 1.012(0.222) 1.029(0.236) 0.982(0.183) 1.007(0.208)
#>   6  0.987(0.211) 1.001(0.205) 0.985(0.166) 0.987(0.211) 1.035(0.194)
#>   7  1.009(0.175) 0.999(0.170) 1.030(0.187) 1.004(0.193) 1.030(0.204)
#>   8  1.004(0.205) 0.987(0.164) 1.006(0.202) 1.008(0.189) 1.005(0.204)
#>   9  0.965(0.158) 1.034(0.257) 1.009(0.178) 0.971(0.179) 1.016(0.213)
#>   10 0.995(0.172) 1.056(0.237) 0.988(0.212) 0.987(0.194) 1.043(0.224)
#>     year
#> age  1985         1986         1987         1988         1989        
#>   1  0.984(0.180) 0.982(0.203) 0.992(0.182) 0.991(0.211) 1.000(0.169)
#>   2  0.979(0.196) 0.972(0.218) 0.998(0.222) 1.036(0.209) 1.017(0.218)
#>   3  1.002(0.225) 1.040(0.218) 1.027(0.172) 1.036(0.178) 1.022(0.187)
#>   4  0.966(0.213) 1.012(0.200) 0.998(0.165) 0.983(0.185) 0.998(0.215)
#>   5  0.963(0.180) 1.014(0.189) 0.983(0.167) 0.993(0.166) 0.999(0.215)
#>   6  1.017(0.194) 0.998(0.189) 0.945(0.175) 1.020(0.179) 1.040(0.238)
#>   7  1.033(0.214) 0.998(0.158) 0.973(0.205) 1.004(0.201) 1.022(0.182)
#>   8  0.976(0.165) 0.997(0.191) 1.000(0.207) 1.003(0.249) 1.011(0.198)
#>   9  1.005(0.228) 0.987(0.188) 1.007(0.240) 1.025(0.229) 0.982(0.220)
#>   10 1.034(0.193) 0.999(0.186) 0.999(0.186) 0.984(0.214) 0.999(0.195)
#>     year
#> age  1990         1991         1992         1993         1994        
#>   1  0.969(0.206) 0.992(0.203) 0.989(0.184) 1.006(0.210) 1.023(0.200)
#>   2  1.049(0.219) 0.997(0.171) 0.979(0.227) 1.027(0.196) 1.032(0.167)
#>   3  1.017(0.187) 1.029(0.240) 0.960(0.211) 0.966(0.178) 0.998(0.170)
#>   4  0.986(0.207) 1.045(0.176) 1.006(0.190) 1.005(0.183) 0.998(0.215)
#>   5  1.015(0.224) 1.017(0.182) 1.039(0.195) 1.014(0.213) 1.001(0.208)
#>   6  1.002(0.173) 1.002(0.156) 1.014(0.204) 0.976(0.212) 0.979(0.208)
#>   7  1.000(0.162) 0.998(0.179) 0.981(0.205) 1.010(0.214) 1.005(0.208)
#>   8  1.050(0.230) 1.034(0.185) 0.983(0.157) 1.043(0.204) 1.017(0.217)
#>   9  0.974(0.203) 1.021(0.175) 0.952(0.203) 1.025(0.232) 0.975(0.175)
#>   10 1.015(0.180) 0.995(0.188) 1.021(0.172) 0.982(0.207) 1.014(0.237)
#>     year
#> age  1995         1996         1997         1998         1999        
#>   1  0.986(0.252) 1.062(0.240) 1.007(0.176) 0.965(0.191) 1.037(0.184)
#>   2  1.028(0.217) 0.994(0.211) 0.985(0.234) 0.983(0.205) 0.998(0.175)
#>   3  0.993(0.190) 0.975(0.172) 0.974(0.213) 0.990(0.181) 1.053(0.210)
#>   4  1.014(0.182) 1.045(0.225) 1.005(0.175) 1.046(0.207) 0.977(0.193)
#>   5  1.029(0.223) 0.937(0.202) 0.998(0.184) 1.056(0.188) 1.028(0.176)
#>   6  1.002(0.204) 0.966(0.174) 0.990(0.161) 0.987(0.200) 0.978(0.175)
#>   7  0.991(0.174) 1.027(0.161) 0.962(0.197) 0.972(0.209) 1.005(0.181)
#>   8  0.985(0.177) 0.973(0.181) 0.997(0.188) 1.003(0.194) 1.008(0.211)
#>   9  0.999(0.208) 0.988(0.212) 0.963(0.215) 1.030(0.179) 0.986(0.185)
#>   10 0.982(0.181) 1.027(0.198) 1.051(0.191) 0.999(0.212) 1.014(0.167)
#>     year
#> age  2000         2001         2002         2003         2004        
#>   1  0.978(0.190) 1.014(0.176) 1.018(0.216) 0.976(0.204) 1.007(0.223)
#>   2  0.989(0.192) 1.046(0.157) 1.009(0.195) 0.969(0.143) 0.995(0.159)
#>   3  0.999(0.198) 0.995(0.195) 1.023(0.197) 0.993(0.213) 1.014(0.213)
#>   4  1.021(0.139) 0.987(0.147) 0.973(0.234) 1.044(0.238) 1.019(0.187)
#>   5  1.021(0.204) 0.953(0.172) 1.027(0.223) 1.019(0.218) 1.029(0.211)
#>   6  1.037(0.185) 0.984(0.233) 0.973(0.197) 1.031(0.235) 1.003(0.196)
#>   7  0.979(0.226) 1.011(0.165) 0.977(0.214) 1.037(0.237) 1.029(0.191)
#>   8  1.014(0.190) 0.956(0.214) 1.015(0.217) 0.986(0.172) 0.986(0.207)
#>   9  1.023(0.206) 0.987(0.162) 0.980(0.184) 1.028(0.235) 1.076(0.260)
#>   10 0.995(0.252) 0.964(0.192) 1.019(0.175) 0.993(0.208) 1.006(0.220)
#>     year
#> age  2005         2006         2007         2008         2009        
#>   1  0.987(0.224) 1.023(0.197) 0.964(0.164) 0.997(0.239) 1.021(0.187)
#>   2  1.036(0.211) 1.045(0.205) 0.983(0.185) 1.045(0.215) 1.022(0.236)
#>   3  1.046(0.192) 1.008(0.181) 1.007(0.191) 1.015(0.189) 0.991(0.203)
#>   4  1.001(0.217) 1.006(0.191) 1.022(0.223) 0.974(0.240) 1.053(0.234)
#>   5  0.974(0.148) 0.978(0.171) 0.991(0.191) 0.997(0.211) 0.990(0.225)
#>   6  1.010(0.205) 1.012(0.197) 1.048(0.183) 0.999(0.224) 1.011(0.181)
#>   7  0.993(0.201) 0.987(0.189) 1.034(0.165) 0.977(0.181) 1.001(0.244)
#>   8  1.000(0.187) 0.965(0.194) 1.000(0.200) 1.013(0.203) 1.007(0.187)
#>   9  0.994(0.198) 1.021(0.184) 0.983(0.171) 1.065(0.219) 1.041(0.162)
#>   10 0.948(0.193) 1.005(0.170) 1.003(0.213) 1.003(0.202) 1.032(0.202)
#>     year
#> age  2010         2011         2012         2013         2014        
#>   1  1.016(0.197) 0.957(0.194) 0.976(0.202) 1.014(0.201) 0.992(0.163)
#>   2  0.996(0.222) 0.991(0.183) 1.026(0.184) 1.028(0.212) 1.036(0.173)
#>   3  0.988(0.196) 1.012(0.179) 0.998(0.179) 0.998(0.179) 0.980(0.193)
#>   4  0.978(0.163) 0.983(0.207) 1.020(0.183) 0.987(0.185) 1.004(0.204)
#>   5  0.999(0.233) 1.047(0.224) 1.005(0.197) 0.975(0.193) 0.988(0.203)
#>   6  0.976(0.164) 1.020(0.167) 1.010(0.202) 1.014(0.223) 0.963(0.197)
#>   7  1.002(0.159) 1.000(0.192) 1.008(0.216) 0.995(0.221) 1.002(0.191)
#>   8  0.963(0.195) 1.036(0.186) 0.993(0.197) 0.973(0.230) 0.972(0.215)
#>   9  0.961(0.183) 1.002(0.200) 0.989(0.186) 1.028(0.200) 1.010(0.234)
#>   10 0.994(0.177) 0.971(0.166) 1.030(0.233) 0.989(0.167) 0.960(0.167)
#>     year
#> age  2015         2016         2017         2018         2019        
#>   1  1.031(0.196) 1.000(0.222) 0.999(0.215) 0.996(0.194) 0.971(0.184)
#>   2  0.999(0.181) 0.973(0.208) 1.019(0.205) 0.966(0.181) 1.019(0.205)
#>   3  0.972(0.207) 0.976(0.167) 1.026(0.225) 0.986(0.190) 0.990(0.162)
#>   4  0.984(0.189) 1.023(0.229) 1.043(0.226) 1.038(0.187) 1.026(0.183)
#>   5  0.942(0.237) 0.989(0.144) 1.014(0.218) 1.044(0.225) 1.017(0.171)
#>   6  0.996(0.244) 0.974(0.182) 1.021(0.192) 0.995(0.208) 0.963(0.192)
#>   7  1.013(0.186) 1.016(0.183) 0.979(0.189) 1.001(0.224) 1.004(0.234)
#>   8  1.018(0.200) 0.960(0.192) 1.009(0.173) 1.027(0.205) 0.999(0.157)
#>   9  1.046(0.225) 0.988(0.174) 1.006(0.215) 1.007(0.195) 1.049(0.184)
#>   10 0.987(0.241) 1.021(0.182) 0.987(0.174) 0.995(0.170) 0.975(0.190)
#>     year
#> age  2020         2021         2022         2023         2024        
#>   1  0.981(0.213) 1.026(0.215) 0.986(0.164) 0.981(0.178) 0.986(0.189)
#>   2  0.996(0.202) 1.019(0.240) 1.026(0.187) 0.976(0.179) 1.009(0.227)
#>   3  0.978(0.240) 1.022(0.206) 1.020(0.185) 0.961(0.193) 1.001(0.176)
#>   4  0.938(0.153) 1.039(0.212) 1.004(0.131) 1.012(0.188) 0.970(0.186)
#>   5  0.946(0.184) 1.006(0.245) 1.019(0.205) 0.978(0.190) 0.999(0.190)
#>   6  1.030(0.205) 1.032(0.229) 0.962(0.177) 1.026(0.202) 0.982(0.178)
#>   7  0.976(0.182) 0.983(0.200) 0.981(0.219) 1.012(0.194) 0.977(0.194)
#>   8  1.027(0.195) 0.960(0.184) 1.015(0.217) 0.970(0.191) 1.026(0.192)
#>   9  0.992(0.208) 0.974(0.190) 0.998(0.180) 0.995(0.191) 1.025(0.246)
#>   10 0.976(0.199) 1.042(0.211) 0.981(0.196) 0.996(0.156) 0.999(0.206)
#>     year
#> age  2025         2026         2027         2028         2029        
#>   1  1.010(0.204) 1.012(0.196) 1.002(0.190) 1.020(0.233) 1.001(0.204)
#>   2  1.044(0.183) 1.008(0.225) 1.009(0.175) 1.009(0.203) 1.019(0.202)
#>   3  0.990(0.190) 1.024(0.203) 0.979(0.155) 1.001(0.231) 0.973(0.198)
#>   4  0.996(0.214) 1.063(0.258) 0.990(0.208) 0.959(0.172) 1.023(0.195)
#>   5  1.023(0.219) 0.991(0.218) 0.976(0.220) 0.981(0.217) 0.965(0.207)
#>   6  1.049(0.226) 0.995(0.201) 0.961(0.206) 0.983(0.198) 1.033(0.227)
#>   7  0.984(0.197) 1.025(0.170) 0.973(0.191) 1.025(0.225) 1.026(0.189)
#>   8  0.996(0.190) 0.994(0.173) 0.999(0.244) 0.982(0.197) 0.998(0.196)
#>   9  0.997(0.178) 1.017(0.246) 0.981(0.180) 0.970(0.199) 1.025(0.239)
#>   10 1.006(0.206) 1.036(0.246) 0.947(0.185) 1.013(0.226) 1.024(0.203)
#>     year
#> age  2030         2031         2032         2033         2034        
#>   1  0.984(0.228) 0.988(0.183) 1.003(0.257) 1.021(0.206) 0.939(0.163)
#>   2  1.001(0.231) 0.952(0.176) 1.043(0.220) 1.004(0.200) 1.012(0.222)
#>   3  1.028(0.223) 0.972(0.200) 1.004(0.219) 0.980(0.188) 0.964(0.198)
#>   4  1.008(0.200) 0.992(0.188) 1.025(0.187) 0.921(0.208) 0.953(0.159)
#>   5  0.973(0.225) 0.996(0.255) 1.003(0.229) 0.999(0.170) 0.985(0.191)
#>   6  1.004(0.161) 1.030(0.210) 1.002(0.199) 0.964(0.155) 0.990(0.216)
#>   7  1.000(0.188) 1.016(0.196) 0.994(0.198) 1.004(0.189) 0.976(0.215)
#>   8  0.995(0.183) 0.988(0.183) 1.020(0.181) 0.995(0.204) 1.016(0.256)
#>   9  1.008(0.200) 1.004(0.160) 1.017(0.173) 1.053(0.201) 1.019(0.207)
#>   10 1.016(0.221) 1.010(0.197) 0.996(0.194) 0.982(0.190) 0.947(0.189)
#>     year
#> age  2035         2036         2037         2038         2039        
#>   1  0.976(0.179) 1.043(0.152) 1.033(0.190) 1.008(0.152) 1.000(0.189)
#>   2  0.942(0.137) 1.019(0.206) 0.984(0.183) 0.990(0.157) 0.988(0.172)
#>   3  0.989(0.181) 1.010(0.230) 0.977(0.207) 0.953(0.180) 0.995(0.213)
#>   4  0.985(0.205) 1.002(0.163) 0.984(0.190) 1.015(0.177) 1.007(0.173)
#>   5  1.033(0.222) 1.015(0.200) 1.013(0.194) 1.000(0.192) 0.984(0.213)
#>   6  0.994(0.246) 0.980(0.167) 0.986(0.178) 1.007(0.176) 0.990(0.179)
#>   7  0.994(0.193) 0.992(0.198) 1.054(0.221) 1.002(0.185) 1.004(0.194)
#>   8  1.039(0.197) 0.959(0.208) 0.971(0.187) 0.995(0.171) 1.021(0.188)
#>   9  1.022(0.202) 0.965(0.198) 0.977(0.222) 0.972(0.195) 0.977(0.193)
#>   10 0.987(0.173) 1.041(0.210) 0.997(0.180) 0.985(0.173) 0.980(0.173)
#>     year
#> age  2040         2041         2042         2043         2044        
#>   1  0.985(0.196) 0.957(0.197) 0.995(0.213) 1.016(0.176) 1.003(0.203)
#>   2  1.000(0.260) 1.011(0.175) 0.980(0.174) 1.011(0.220) 1.008(0.172)
#>   3  1.034(0.216) 0.986(0.171) 1.026(0.194) 1.038(0.199) 1.046(0.189)
#>   4  1.029(0.228) 1.003(0.177) 1.011(0.214) 1.022(0.193) 1.008(0.214)
#>   5  1.011(0.201) 1.006(0.186) 1.026(0.212) 0.993(0.191) 1.007(0.126)
#>   6  1.016(0.194) 1.020(0.232) 0.976(0.203) 0.988(0.221) 0.993(0.188)
#>   7  1.018(0.232) 1.016(0.234) 1.005(0.165) 0.965(0.184) 0.957(0.251)
#>   8  0.976(0.239) 0.997(0.173) 1.024(0.216) 0.960(0.158) 0.990(0.170)
#>   9  0.963(0.170) 0.972(0.225) 1.043(0.203) 1.019(0.156) 0.979(0.220)
#>   10 1.063(0.200) 1.002(0.169) 1.013(0.228) 1.001(0.219) 1.040(0.180)
#>     year
#> age  2045         2046         2047         2048         2049        
#>   1  0.992(0.181) 1.050(0.186) 1.001(0.206) 0.946(0.189) 1.052(0.203)
#>   2  1.016(0.236) 0.985(0.175) 1.011(0.246) 0.973(0.170) 1.017(0.167)
#>   3  0.981(0.193) 0.982(0.187) 1.024(0.175) 1.019(0.223) 1.015(0.170)
#>   4  1.011(0.199) 1.014(0.228) 1.053(0.201) 1.010(0.210) 1.005(0.191)
#>   5  1.003(0.199) 1.010(0.216) 1.009(0.166) 0.976(0.193) 0.993(0.214)
#>   6  1.003(0.212) 1.007(0.193) 0.971(0.181) 1.022(0.205) 0.969(0.191)
#>   7  1.009(0.171) 1.006(0.178) 1.009(0.182) 0.994(0.194) 1.018(0.170)
#>   8  0.991(0.222) 1.012(0.215) 1.015(0.196) 1.018(0.206) 0.985(0.217)
#>   9  1.016(0.214) 0.987(0.196) 1.003(0.185) 1.009(0.195) 0.981(0.195)
#>   10 1.022(0.254) 1.058(0.200) 1.009(0.238) 1.012(0.196) 1.000(0.197)
#>     year
#> age  2050         2051         2052         2053         2054        
#>   1  0.983(0.195) 1.001(0.222) 1.010(0.205) 1.016(0.234) 1.041(0.233)
#>   2  1.044(0.237) 1.021(0.217) 0.985(0.214) 0.997(0.194) 0.956(0.198)
#>   3  0.984(0.160) 1.007(0.217) 1.067(0.194) 1.033(0.254) 1.008(0.238)
#>   4  0.968(0.183) 0.994(0.183) 1.007(0.200) 0.968(0.233) 0.996(0.194)
#>   5  0.957(0.202) 1.011(0.179) 1.001(0.183) 0.983(0.194) 0.989(0.211)
#>   6  0.980(0.176) 1.001(0.218) 1.007(0.173) 0.979(0.166) 0.953(0.161)
#>   7  1.057(0.214) 1.003(0.206) 0.967(0.176) 0.988(0.185) 1.020(0.204)
#>   8  0.997(0.152) 0.999(0.219) 1.043(0.229) 0.977(0.226) 1.014(0.194)
#>   9  1.000(0.203) 0.980(0.178) 0.965(0.162) 0.986(0.166) 0.999(0.191)
#>   10 0.976(0.190) 1.016(0.187) 1.026(0.196) 0.963(0.182) 0.979(0.201)
#>     year
#> age  2055        
#>   1  1.018(0.161)
#>   2  0.994(0.163)
#>   3  0.945(0.192)
#>   4  0.944(0.172)
#>   5  1.030(0.180)
#>   6  0.994(0.174)
#>   7  1.022(0.212)
#>   8  0.990(0.212)
#>   9  1.000(0.197)
#>   10 1.011(0.188)
#> 
#> units:  1000 
#> 
#> 
#> $idx
#> $ SUR 
#> An object of class "FLQuant"
#> iters:  100 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1980         1981         1982         1983         1984        
#>   all 0.999(0.173) 0.971(0.220) 0.975(0.211) 0.964(0.172) 0.976(0.182)
#>      year
#> quant 1985         1986         1987         1988         1989        
#>   all 1.017(0.205) 0.976(0.158) 1.005(0.154) 1.016(0.199) 0.980(0.197)
#>      year
#> quant 1990         1991         1992         1993         1994        
#>   all 1.015(0.207) 0.958(0.178) 1.029(0.214) 0.962(0.183) 0.961(0.210)
#>      year
#> quant 1995         1996         1997         1998         1999        
#>   all 0.989(0.201) 1.050(0.173) 1.076(0.241) 0.991(0.189) 1.020(0.184)
#>      year
#> quant 2000         2001         2002         2003         2004        
#>   all 0.974(0.173) 1.006(0.232) 1.014(0.253) 1.022(0.210) 1.015(0.164)
#>      year
#> quant 2005         2006         2007         2008         2009        
#>   all 0.995(0.175) 1.039(0.235) 0.966(0.168) 0.989(0.219) 0.969(0.190)
#>      year
#> quant 2010         2011         2012         2013         2014        
#>   all 0.983(0.212) 0.997(0.174) 1.041(0.186) 1.000(0.254) 1.006(0.189)
#>      year
#> quant 2015         2016         2017         2018         2019        
#>   all 1.003(0.221) 1.019(0.233) 0.998(0.228) 0.977(0.250) 0.968(0.202)
#>      year
#> quant 2020         2021         2022         2023         2024        
#>   all 0.983(0.203) 0.976(0.190) 0.966(0.200) 0.967(0.210) 1.022(0.238)
#>      year
#> quant 2025         2026         2027         2028         2029        
#>   all 0.988(0.168) 0.984(0.214) 1.014(0.204) 1.002(0.225) 1.021(0.196)
#>      year
#> quant 2030         2031         2032         2033         2034        
#>   all 1.012(0.217) 0.995(0.185) 1.019(0.193) 0.985(0.242) 1.012(0.196)
#>      year
#> quant 2035         2036         2037         2038         2039        
#>   all 0.993(0.197) 0.998(0.203) 0.995(0.176) 0.976(0.176) 1.070(0.181)
#>      year
#> quant 2040         2041         2042         2043         2044        
#>   all 0.973(0.185) 1.035(0.228) 0.988(0.186) 0.960(0.192) 0.996(0.179)
#>      year
#> quant 2045         2046         2047         2048         2049        
#>   all 1.028(0.198) 1.009(0.202) 1.010(0.215) 1.026(0.188) 1.051(0.188)
#>      year
#> quant 2050         2051         2052         2053         2054        
#>   all 0.988(0.186) 1.002(0.194) 1.016(0.186) 0.990(0.198) 1.003(0.186)
#>      year
#> quant 2055        
#>   all 1.026(0.225)
#> 
#> units:  NA 
#> 
#> $ CPUE 
#> An object of class "FLQuant"
#> iters:  100 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1980         1981         1982         1983         1984        
#>   all 0.996(0.251) 1.042(0.290) 1.033(0.302) 1.010(0.238) 1.021(0.313)
#>      year
#> quant 1985         1986         1987         1988         1989        
#>   all 0.952(0.240) 1.009(0.265) 0.958(0.333) 0.993(0.299) 0.958(0.275)
#>      year
#> quant 1990         1991         1992         1993         1994        
#>   all 0.983(0.299) 1.021(0.347) 1.044(0.284) 0.914(0.251) 0.951(0.257)
#>      year
#> quant 1995         1996         1997         1998         1999        
#>   all 0.975(0.328) 0.925(0.325) 0.992(0.291) 1.020(0.317) 0.951(0.260)
#>      year
#> quant 2000         2001         2002         2003         2004        
#>   all 1.004(0.253) 0.990(0.235) 1.044(0.345) 0.939(0.292) 1.020(0.300)
#>      year
#> quant 2005         2006         2007         2008         2009        
#>   all 1.044(0.276) 1.017(0.280) 0.952(0.278) 0.931(0.274) 1.007(0.226)
#>      year
#> quant 2010         2011         2012         2013         2014        
#>   all 1.068(0.299) 0.979(0.304) 1.010(0.353) 1.037(0.342) 0.953(0.296)
#>      year
#> quant 2015         2016         2017         2018         2019        
#>   all 0.977(0.313) 0.983(0.215) 0.942(0.305) 0.993(0.233) 1.026(0.268)
#>      year
#> quant 2020         2021         2022         2023         2024        
#>   all 0.965(0.263) 1.006(0.312) 0.988(0.346) 0.933(0.303) 0.998(0.291)
#>      year
#> quant 2025         2026         2027         2028         2029        
#>   all 1.016(0.269) 0.976(0.299) 1.012(0.244) 0.969(0.257) 1.009(0.356)
#>      year
#> quant 2030         2031         2032         2033         2034        
#>   all 1.007(0.316) 0.969(0.269) 0.999(0.347) 0.964(0.312) 0.994(0.290)
#>      year
#> quant 2035         2036         2037         2038         2039        
#>   all 0.972(0.250) 1.052(0.362) 0.990(0.272) 0.976(0.230) 1.025(0.291)
#>      year
#> quant 2040         2041         2042         2043         2044        
#>   all 0.972(0.317) 0.981(0.317) 1.021(0.324) 0.994(0.298) 0.978(0.241)
#>      year
#> quant 2045         2046         2047         2048         2049        
#>   all 1.048(0.285) 0.965(0.336) 1.045(0.333) 1.016(0.338) 1.087(0.357)
#>      year
#> quant 2050         2051         2052         2053         2054        
#>   all 1.094(0.352) 1.030(0.287) 1.000(0.344) 0.993(0.308) 1.015(0.302)
#>      year
#> quant 2055        
#>   all 1.073(0.256)
#> 
#> units:  NA 
#> 
#> 
deviances(oem, "stk")
#> $ catch.n 
#> An object of class "FLQuant"
#> iters:  100 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1960         1961         1962         1963         1964        
#>   1  0.997(0.176) 0.976(0.191) 1.098(0.189) 0.984(0.220) 1.015(0.190)
#>   2  1.002(0.202) 1.013(0.220) 1.008(0.193) 0.964(0.196) 0.988(0.220)
#>   3  1.001(0.189) 0.947(0.175) 0.961(0.169) 0.972(0.254) 1.036(0.210)
#>   4  1.019(0.184) 0.999(0.196) 1.045(0.275) 1.014(0.189) 1.022(0.168)
#>   5  1.048(0.194) 0.992(0.210) 0.987(0.206) 1.035(0.198) 1.000(0.213)
#>   6  0.980(0.179) 0.989(0.174) 0.957(0.196) 1.026(0.215) 0.979(0.168)
#>   7  0.995(0.219) 0.984(0.186) 1.026(0.211) 0.979(0.201) 1.014(0.201)
#>   8  1.036(0.186) 1.019(0.211) 1.019(0.217) 1.030(0.208) 0.956(0.153)
#>   9  0.991(0.220) 0.992(0.173) 1.012(0.222) 0.998(0.249) 0.968(0.211)
#>   10 1.015(0.185) 0.989(0.210) 1.005(0.165) 0.990(0.222) 0.958(0.207)
#>     year
#> age  1965         1966         1967         1968         1969        
#>   1  0.962(0.206) 1.026(0.205) 1.020(0.228) 0.994(0.223) 0.987(0.169)
#>   2  0.994(0.186) 1.038(0.174) 0.960(0.216) 0.971(0.171) 1.023(0.258)
#>   3  1.020(0.196) 0.977(0.214) 0.993(0.200) 0.985(0.190) 1.007(0.208)
#>   4  1.018(0.183) 1.029(0.208) 0.979(0.207) 1.013(0.184) 0.990(0.194)
#>   5  0.997(0.168) 0.973(0.198) 1.004(0.186) 0.995(0.208) 0.993(0.185)
#>   6  0.981(0.200) 0.956(0.172) 0.975(0.174) 0.970(0.190) 1.021(0.230)
#>   7  1.001(0.159) 0.966(0.190) 1.006(0.211) 0.974(0.183) 1.017(0.177)
#>   8  0.976(0.231) 0.970(0.135) 1.035(0.186) 1.004(0.201) 1.013(0.202)
#>   9  1.031(0.276) 1.001(0.183) 1.025(0.187) 0.971(0.203) 1.035(0.184)
#>   10 1.013(0.180) 0.993(0.156) 1.017(0.198) 0.953(0.208) 0.996(0.189)
#>     year
#> age  1970         1971         1972         1973         1974        
#>   1  1.002(0.174) 0.965(0.177) 0.996(0.153) 1.013(0.178) 1.012(0.202)
#>   2  0.998(0.173) 0.966(0.182) 0.984(0.182) 1.008(0.171) 0.983(0.209)
#>   3  0.993(0.239) 0.985(0.206) 0.995(0.213) 1.013(0.209) 1.008(0.191)
#>   4  0.979(0.232) 0.998(0.185) 1.026(0.211) 0.966(0.217) 0.970(0.205)
#>   5  1.008(0.182) 0.997(0.236) 1.009(0.144) 1.021(0.184) 0.982(0.188)
#>   6  0.976(0.191) 1.004(0.181) 0.963(0.196) 0.961(0.187) 0.968(0.183)
#>   7  1.004(0.192) 1.019(0.208) 0.954(0.175) 0.992(0.197) 1.010(0.231)
#>   8  0.948(0.187) 0.960(0.224) 0.992(0.152) 0.988(0.221) 0.989(0.177)
#>   9  1.030(0.203) 0.975(0.162) 0.997(0.186) 0.984(0.184) 1.015(0.222)
#>   10 1.038(0.182) 0.995(0.185) 0.982(0.210) 1.003(0.220) 0.976(0.186)
#>     year
#> age  1975         1976         1977         1978         1979        
#>   1  0.958(0.216) 0.999(0.192) 0.971(0.230) 1.030(0.184) 0.984(0.203)
#>   2  0.991(0.213) 1.023(0.226) 0.996(0.235) 0.985(0.197) 1.018(0.176)
#>   3  1.029(0.207) 1.015(0.197) 1.017(0.189) 1.004(0.224) 0.998(0.195)
#>   4  1.006(0.205) 1.004(0.175) 1.013(0.222) 0.983(0.193) 0.985(0.210)
#>   5  1.056(0.218) 0.971(0.175) 0.973(0.244) 1.018(0.205) 1.042(0.188)
#>   6  0.992(0.200) 0.978(0.202) 1.050(0.219) 1.039(0.193) 0.970(0.201)
#>   7  0.967(0.172) 0.983(0.197) 0.989(0.203) 0.969(0.197) 1.030(0.200)
#>   8  0.997(0.160) 0.990(0.183) 0.970(0.191) 0.984(0.203) 1.013(0.242)
#>   9  0.981(0.193) 0.976(0.182) 0.981(0.208) 1.042(0.219) 1.020(0.176)
#>   10 0.999(0.181) 0.976(0.194) 0.965(0.185) 0.963(0.218) 1.003(0.189)
#>     year
#> age  1980         1981         1982         1983         1984        
#>   1  1.082(0.209) 0.994(0.217) 1.076(0.224) 0.997(0.217) 0.985(0.185)
#>   2  1.007(0.183) 0.997(0.200) 0.991(0.147) 1.041(0.240) 0.956(0.200)
#>   3  1.011(0.212) 1.021(0.162) 1.001(0.172) 1.001(0.200) 1.025(0.160)
#>   4  0.993(0.218) 0.987(0.168) 1.000(0.185) 0.969(0.221) 1.024(0.232)
#>   5  0.990(0.159) 1.012(0.222) 1.029(0.236) 0.982(0.183) 1.007(0.208)
#>   6  0.987(0.211) 1.001(0.205) 0.985(0.166) 0.987(0.211) 1.035(0.194)
#>   7  1.009(0.175) 0.999(0.170) 1.030(0.187) 1.004(0.193) 1.030(0.204)
#>   8  1.004(0.205) 0.987(0.164) 1.006(0.202) 1.008(0.189) 1.005(0.204)
#>   9  0.965(0.158) 1.034(0.257) 1.009(0.178) 0.971(0.179) 1.016(0.213)
#>   10 0.995(0.172) 1.056(0.237) 0.988(0.212) 0.987(0.194) 1.043(0.224)
#>     year
#> age  1985         1986         1987         1988         1989        
#>   1  0.984(0.180) 0.982(0.203) 0.992(0.182) 0.991(0.211) 1.000(0.169)
#>   2  0.979(0.196) 0.972(0.218) 0.998(0.222) 1.036(0.209) 1.017(0.218)
#>   3  1.002(0.225) 1.040(0.218) 1.027(0.172) 1.036(0.178) 1.022(0.187)
#>   4  0.966(0.213) 1.012(0.200) 0.998(0.165) 0.983(0.185) 0.998(0.215)
#>   5  0.963(0.180) 1.014(0.189) 0.983(0.167) 0.993(0.166) 0.999(0.215)
#>   6  1.017(0.194) 0.998(0.189) 0.945(0.175) 1.020(0.179) 1.040(0.238)
#>   7  1.033(0.214) 0.998(0.158) 0.973(0.205) 1.004(0.201) 1.022(0.182)
#>   8  0.976(0.165) 0.997(0.191) 1.000(0.207) 1.003(0.249) 1.011(0.198)
#>   9  1.005(0.228) 0.987(0.188) 1.007(0.240) 1.025(0.229) 0.982(0.220)
#>   10 1.034(0.193) 0.999(0.186) 0.999(0.186) 0.984(0.214) 0.999(0.195)
#>     year
#> age  1990         1991         1992         1993         1994        
#>   1  0.969(0.206) 0.992(0.203) 0.989(0.184) 1.006(0.210) 1.023(0.200)
#>   2  1.049(0.219) 0.997(0.171) 0.979(0.227) 1.027(0.196) 1.032(0.167)
#>   3  1.017(0.187) 1.029(0.240) 0.960(0.211) 0.966(0.178) 0.998(0.170)
#>   4  0.986(0.207) 1.045(0.176) 1.006(0.190) 1.005(0.183) 0.998(0.215)
#>   5  1.015(0.224) 1.017(0.182) 1.039(0.195) 1.014(0.213) 1.001(0.208)
#>   6  1.002(0.173) 1.002(0.156) 1.014(0.204) 0.976(0.212) 0.979(0.208)
#>   7  1.000(0.162) 0.998(0.179) 0.981(0.205) 1.010(0.214) 1.005(0.208)
#>   8  1.050(0.230) 1.034(0.185) 0.983(0.157) 1.043(0.204) 1.017(0.217)
#>   9  0.974(0.203) 1.021(0.175) 0.952(0.203) 1.025(0.232) 0.975(0.175)
#>   10 1.015(0.180) 0.995(0.188) 1.021(0.172) 0.982(0.207) 1.014(0.237)
#>     year
#> age  1995         1996         1997         1998         1999        
#>   1  0.986(0.252) 1.062(0.240) 1.007(0.176) 0.965(0.191) 1.037(0.184)
#>   2  1.028(0.217) 0.994(0.211) 0.985(0.234) 0.983(0.205) 0.998(0.175)
#>   3  0.993(0.190) 0.975(0.172) 0.974(0.213) 0.990(0.181) 1.053(0.210)
#>   4  1.014(0.182) 1.045(0.225) 1.005(0.175) 1.046(0.207) 0.977(0.193)
#>   5  1.029(0.223) 0.937(0.202) 0.998(0.184) 1.056(0.188) 1.028(0.176)
#>   6  1.002(0.204) 0.966(0.174) 0.990(0.161) 0.987(0.200) 0.978(0.175)
#>   7  0.991(0.174) 1.027(0.161) 0.962(0.197) 0.972(0.209) 1.005(0.181)
#>   8  0.985(0.177) 0.973(0.181) 0.997(0.188) 1.003(0.194) 1.008(0.211)
#>   9  0.999(0.208) 0.988(0.212) 0.963(0.215) 1.030(0.179) 0.986(0.185)
#>   10 0.982(0.181) 1.027(0.198) 1.051(0.191) 0.999(0.212) 1.014(0.167)
#>     year
#> age  2000         2001         2002         2003         2004        
#>   1  0.978(0.190) 1.014(0.176) 1.018(0.216) 0.976(0.204) 1.007(0.223)
#>   2  0.989(0.192) 1.046(0.157) 1.009(0.195) 0.969(0.143) 0.995(0.159)
#>   3  0.999(0.198) 0.995(0.195) 1.023(0.197) 0.993(0.213) 1.014(0.213)
#>   4  1.021(0.139) 0.987(0.147) 0.973(0.234) 1.044(0.238) 1.019(0.187)
#>   5  1.021(0.204) 0.953(0.172) 1.027(0.223) 1.019(0.218) 1.029(0.211)
#>   6  1.037(0.185) 0.984(0.233) 0.973(0.197) 1.031(0.235) 1.003(0.196)
#>   7  0.979(0.226) 1.011(0.165) 0.977(0.214) 1.037(0.237) 1.029(0.191)
#>   8  1.014(0.190) 0.956(0.214) 1.015(0.217) 0.986(0.172) 0.986(0.207)
#>   9  1.023(0.206) 0.987(0.162) 0.980(0.184) 1.028(0.235) 1.076(0.260)
#>   10 0.995(0.252) 0.964(0.192) 1.019(0.175) 0.993(0.208) 1.006(0.220)
#>     year
#> age  2005         2006         2007         2008         2009        
#>   1  0.987(0.224) 1.023(0.197) 0.964(0.164) 0.997(0.239) 1.021(0.187)
#>   2  1.036(0.211) 1.045(0.205) 0.983(0.185) 1.045(0.215) 1.022(0.236)
#>   3  1.046(0.192) 1.008(0.181) 1.007(0.191) 1.015(0.189) 0.991(0.203)
#>   4  1.001(0.217) 1.006(0.191) 1.022(0.223) 0.974(0.240) 1.053(0.234)
#>   5  0.974(0.148) 0.978(0.171) 0.991(0.191) 0.997(0.211) 0.990(0.225)
#>   6  1.010(0.205) 1.012(0.197) 1.048(0.183) 0.999(0.224) 1.011(0.181)
#>   7  0.993(0.201) 0.987(0.189) 1.034(0.165) 0.977(0.181) 1.001(0.244)
#>   8  1.000(0.187) 0.965(0.194) 1.000(0.200) 1.013(0.203) 1.007(0.187)
#>   9  0.994(0.198) 1.021(0.184) 0.983(0.171) 1.065(0.219) 1.041(0.162)
#>   10 0.948(0.193) 1.005(0.170) 1.003(0.213) 1.003(0.202) 1.032(0.202)
#>     year
#> age  2010         2011         2012         2013         2014        
#>   1  1.016(0.197) 0.957(0.194) 0.976(0.202) 1.014(0.201) 0.992(0.163)
#>   2  0.996(0.222) 0.991(0.183) 1.026(0.184) 1.028(0.212) 1.036(0.173)
#>   3  0.988(0.196) 1.012(0.179) 0.998(0.179) 0.998(0.179) 0.980(0.193)
#>   4  0.978(0.163) 0.983(0.207) 1.020(0.183) 0.987(0.185) 1.004(0.204)
#>   5  0.999(0.233) 1.047(0.224) 1.005(0.197) 0.975(0.193) 0.988(0.203)
#>   6  0.976(0.164) 1.020(0.167) 1.010(0.202) 1.014(0.223) 0.963(0.197)
#>   7  1.002(0.159) 1.000(0.192) 1.008(0.216) 0.995(0.221) 1.002(0.191)
#>   8  0.963(0.195) 1.036(0.186) 0.993(0.197) 0.973(0.230) 0.972(0.215)
#>   9  0.961(0.183) 1.002(0.200) 0.989(0.186) 1.028(0.200) 1.010(0.234)
#>   10 0.994(0.177) 0.971(0.166) 1.030(0.233) 0.989(0.167) 0.960(0.167)
#>     year
#> age  2015         2016         2017         2018         2019        
#>   1  1.031(0.196) 1.000(0.222) 0.999(0.215) 0.996(0.194) 0.971(0.184)
#>   2  0.999(0.181) 0.973(0.208) 1.019(0.205) 0.966(0.181) 1.019(0.205)
#>   3  0.972(0.207) 0.976(0.167) 1.026(0.225) 0.986(0.190) 0.990(0.162)
#>   4  0.984(0.189) 1.023(0.229) 1.043(0.226) 1.038(0.187) 1.026(0.183)
#>   5  0.942(0.237) 0.989(0.144) 1.014(0.218) 1.044(0.225) 1.017(0.171)
#>   6  0.996(0.244) 0.974(0.182) 1.021(0.192) 0.995(0.208) 0.963(0.192)
#>   7  1.013(0.186) 1.016(0.183) 0.979(0.189) 1.001(0.224) 1.004(0.234)
#>   8  1.018(0.200) 0.960(0.192) 1.009(0.173) 1.027(0.205) 0.999(0.157)
#>   9  1.046(0.225) 0.988(0.174) 1.006(0.215) 1.007(0.195) 1.049(0.184)
#>   10 0.987(0.241) 1.021(0.182) 0.987(0.174) 0.995(0.170) 0.975(0.190)
#>     year
#> age  2020         2021         2022         2023         2024        
#>   1  0.981(0.213) 1.026(0.215) 0.986(0.164) 0.981(0.178) 0.986(0.189)
#>   2  0.996(0.202) 1.019(0.240) 1.026(0.187) 0.976(0.179) 1.009(0.227)
#>   3  0.978(0.240) 1.022(0.206) 1.020(0.185) 0.961(0.193) 1.001(0.176)
#>   4  0.938(0.153) 1.039(0.212) 1.004(0.131) 1.012(0.188) 0.970(0.186)
#>   5  0.946(0.184) 1.006(0.245) 1.019(0.205) 0.978(0.190) 0.999(0.190)
#>   6  1.030(0.205) 1.032(0.229) 0.962(0.177) 1.026(0.202) 0.982(0.178)
#>   7  0.976(0.182) 0.983(0.200) 0.981(0.219) 1.012(0.194) 0.977(0.194)
#>   8  1.027(0.195) 0.960(0.184) 1.015(0.217) 0.970(0.191) 1.026(0.192)
#>   9  0.992(0.208) 0.974(0.190) 0.998(0.180) 0.995(0.191) 1.025(0.246)
#>   10 0.976(0.199) 1.042(0.211) 0.981(0.196) 0.996(0.156) 0.999(0.206)
#>     year
#> age  2025         2026         2027         2028         2029        
#>   1  1.010(0.204) 1.012(0.196) 1.002(0.190) 1.020(0.233) 1.001(0.204)
#>   2  1.044(0.183) 1.008(0.225) 1.009(0.175) 1.009(0.203) 1.019(0.202)
#>   3  0.990(0.190) 1.024(0.203) 0.979(0.155) 1.001(0.231) 0.973(0.198)
#>   4  0.996(0.214) 1.063(0.258) 0.990(0.208) 0.959(0.172) 1.023(0.195)
#>   5  1.023(0.219) 0.991(0.218) 0.976(0.220) 0.981(0.217) 0.965(0.207)
#>   6  1.049(0.226) 0.995(0.201) 0.961(0.206) 0.983(0.198) 1.033(0.227)
#>   7  0.984(0.197) 1.025(0.170) 0.973(0.191) 1.025(0.225) 1.026(0.189)
#>   8  0.996(0.190) 0.994(0.173) 0.999(0.244) 0.982(0.197) 0.998(0.196)
#>   9  0.997(0.178) 1.017(0.246) 0.981(0.180) 0.970(0.199) 1.025(0.239)
#>   10 1.006(0.206) 1.036(0.246) 0.947(0.185) 1.013(0.226) 1.024(0.203)
#>     year
#> age  2030         2031         2032         2033         2034        
#>   1  0.984(0.228) 0.988(0.183) 1.003(0.257) 1.021(0.206) 0.939(0.163)
#>   2  1.001(0.231) 0.952(0.176) 1.043(0.220) 1.004(0.200) 1.012(0.222)
#>   3  1.028(0.223) 0.972(0.200) 1.004(0.219) 0.980(0.188) 0.964(0.198)
#>   4  1.008(0.200) 0.992(0.188) 1.025(0.187) 0.921(0.208) 0.953(0.159)
#>   5  0.973(0.225) 0.996(0.255) 1.003(0.229) 0.999(0.170) 0.985(0.191)
#>   6  1.004(0.161) 1.030(0.210) 1.002(0.199) 0.964(0.155) 0.990(0.216)
#>   7  1.000(0.188) 1.016(0.196) 0.994(0.198) 1.004(0.189) 0.976(0.215)
#>   8  0.995(0.183) 0.988(0.183) 1.020(0.181) 0.995(0.204) 1.016(0.256)
#>   9  1.008(0.200) 1.004(0.160) 1.017(0.173) 1.053(0.201) 1.019(0.207)
#>   10 1.016(0.221) 1.010(0.197) 0.996(0.194) 0.982(0.190) 0.947(0.189)
#>     year
#> age  2035         2036         2037         2038         2039        
#>   1  0.976(0.179) 1.043(0.152) 1.033(0.190) 1.008(0.152) 1.000(0.189)
#>   2  0.942(0.137) 1.019(0.206) 0.984(0.183) 0.990(0.157) 0.988(0.172)
#>   3  0.989(0.181) 1.010(0.230) 0.977(0.207) 0.953(0.180) 0.995(0.213)
#>   4  0.985(0.205) 1.002(0.163) 0.984(0.190) 1.015(0.177) 1.007(0.173)
#>   5  1.033(0.222) 1.015(0.200) 1.013(0.194) 1.000(0.192) 0.984(0.213)
#>   6  0.994(0.246) 0.980(0.167) 0.986(0.178) 1.007(0.176) 0.990(0.179)
#>   7  0.994(0.193) 0.992(0.198) 1.054(0.221) 1.002(0.185) 1.004(0.194)
#>   8  1.039(0.197) 0.959(0.208) 0.971(0.187) 0.995(0.171) 1.021(0.188)
#>   9  1.022(0.202) 0.965(0.198) 0.977(0.222) 0.972(0.195) 0.977(0.193)
#>   10 0.987(0.173) 1.041(0.210) 0.997(0.180) 0.985(0.173) 0.980(0.173)
#>     year
#> age  2040         2041         2042         2043         2044        
#>   1  0.985(0.196) 0.957(0.197) 0.995(0.213) 1.016(0.176) 1.003(0.203)
#>   2  1.000(0.260) 1.011(0.175) 0.980(0.174) 1.011(0.220) 1.008(0.172)
#>   3  1.034(0.216) 0.986(0.171) 1.026(0.194) 1.038(0.199) 1.046(0.189)
#>   4  1.029(0.228) 1.003(0.177) 1.011(0.214) 1.022(0.193) 1.008(0.214)
#>   5  1.011(0.201) 1.006(0.186) 1.026(0.212) 0.993(0.191) 1.007(0.126)
#>   6  1.016(0.194) 1.020(0.232) 0.976(0.203) 0.988(0.221) 0.993(0.188)
#>   7  1.018(0.232) 1.016(0.234) 1.005(0.165) 0.965(0.184) 0.957(0.251)
#>   8  0.976(0.239) 0.997(0.173) 1.024(0.216) 0.960(0.158) 0.990(0.170)
#>   9  0.963(0.170) 0.972(0.225) 1.043(0.203) 1.019(0.156) 0.979(0.220)
#>   10 1.063(0.200) 1.002(0.169) 1.013(0.228) 1.001(0.219) 1.040(0.180)
#>     year
#> age  2045         2046         2047         2048         2049        
#>   1  0.992(0.181) 1.050(0.186) 1.001(0.206) 0.946(0.189) 1.052(0.203)
#>   2  1.016(0.236) 0.985(0.175) 1.011(0.246) 0.973(0.170) 1.017(0.167)
#>   3  0.981(0.193) 0.982(0.187) 1.024(0.175) 1.019(0.223) 1.015(0.170)
#>   4  1.011(0.199) 1.014(0.228) 1.053(0.201) 1.010(0.210) 1.005(0.191)
#>   5  1.003(0.199) 1.010(0.216) 1.009(0.166) 0.976(0.193) 0.993(0.214)
#>   6  1.003(0.212) 1.007(0.193) 0.971(0.181) 1.022(0.205) 0.969(0.191)
#>   7  1.009(0.171) 1.006(0.178) 1.009(0.182) 0.994(0.194) 1.018(0.170)
#>   8  0.991(0.222) 1.012(0.215) 1.015(0.196) 1.018(0.206) 0.985(0.217)
#>   9  1.016(0.214) 0.987(0.196) 1.003(0.185) 1.009(0.195) 0.981(0.195)
#>   10 1.022(0.254) 1.058(0.200) 1.009(0.238) 1.012(0.196) 1.000(0.197)
#>     year
#> age  2050         2051         2052         2053         2054        
#>   1  0.983(0.195) 1.001(0.222) 1.010(0.205) 1.016(0.234) 1.041(0.233)
#>   2  1.044(0.237) 1.021(0.217) 0.985(0.214) 0.997(0.194) 0.956(0.198)
#>   3  0.984(0.160) 1.007(0.217) 1.067(0.194) 1.033(0.254) 1.008(0.238)
#>   4  0.968(0.183) 0.994(0.183) 1.007(0.200) 0.968(0.233) 0.996(0.194)
#>   5  0.957(0.202) 1.011(0.179) 1.001(0.183) 0.983(0.194) 0.989(0.211)
#>   6  0.980(0.176) 1.001(0.218) 1.007(0.173) 0.979(0.166) 0.953(0.161)
#>   7  1.057(0.214) 1.003(0.206) 0.967(0.176) 0.988(0.185) 1.020(0.204)
#>   8  0.997(0.152) 0.999(0.219) 1.043(0.229) 0.977(0.226) 1.014(0.194)
#>   9  1.000(0.203) 0.980(0.178) 0.965(0.162) 0.986(0.166) 0.999(0.191)
#>   10 0.976(0.190) 1.016(0.187) 1.026(0.196) 0.963(0.182) 0.979(0.201)
#>     year
#> age  2055        
#>   1  1.018(0.161)
#>   2  0.994(0.163)
#>   3  0.945(0.192)
#>   4  0.944(0.172)
#>   5  1.030(0.180)
#>   6  0.994(0.174)
#>   7  1.022(0.212)
#>   8  0.990(0.212)
#>   9  1.000(0.197)
#>   10 1.011(0.188)
#> 
#> units:  1000 
#> 
deviances(oem, "stk", "catch.n")
#> An object of class "FLQuant"
#> iters:  100 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1960         1961         1962         1963         1964        
#>   1  0.997(0.176) 0.976(0.191) 1.098(0.189) 0.984(0.220) 1.015(0.190)
#>   2  1.002(0.202) 1.013(0.220) 1.008(0.193) 0.964(0.196) 0.988(0.220)
#>   3  1.001(0.189) 0.947(0.175) 0.961(0.169) 0.972(0.254) 1.036(0.210)
#>   4  1.019(0.184) 0.999(0.196) 1.045(0.275) 1.014(0.189) 1.022(0.168)
#>   5  1.048(0.194) 0.992(0.210) 0.987(0.206) 1.035(0.198) 1.000(0.213)
#>   6  0.980(0.179) 0.989(0.174) 0.957(0.196) 1.026(0.215) 0.979(0.168)
#>   7  0.995(0.219) 0.984(0.186) 1.026(0.211) 0.979(0.201) 1.014(0.201)
#>   8  1.036(0.186) 1.019(0.211) 1.019(0.217) 1.030(0.208) 0.956(0.153)
#>   9  0.991(0.220) 0.992(0.173) 1.012(0.222) 0.998(0.249) 0.968(0.211)
#>   10 1.015(0.185) 0.989(0.210) 1.005(0.165) 0.990(0.222) 0.958(0.207)
#>     year
#> age  1965         1966         1967         1968         1969        
#>   1  0.962(0.206) 1.026(0.205) 1.020(0.228) 0.994(0.223) 0.987(0.169)
#>   2  0.994(0.186) 1.038(0.174) 0.960(0.216) 0.971(0.171) 1.023(0.258)
#>   3  1.020(0.196) 0.977(0.214) 0.993(0.200) 0.985(0.190) 1.007(0.208)
#>   4  1.018(0.183) 1.029(0.208) 0.979(0.207) 1.013(0.184) 0.990(0.194)
#>   5  0.997(0.168) 0.973(0.198) 1.004(0.186) 0.995(0.208) 0.993(0.185)
#>   6  0.981(0.200) 0.956(0.172) 0.975(0.174) 0.970(0.190) 1.021(0.230)
#>   7  1.001(0.159) 0.966(0.190) 1.006(0.211) 0.974(0.183) 1.017(0.177)
#>   8  0.976(0.231) 0.970(0.135) 1.035(0.186) 1.004(0.201) 1.013(0.202)
#>   9  1.031(0.276) 1.001(0.183) 1.025(0.187) 0.971(0.203) 1.035(0.184)
#>   10 1.013(0.180) 0.993(0.156) 1.017(0.198) 0.953(0.208) 0.996(0.189)
#>     year
#> age  1970         1971         1972         1973         1974        
#>   1  1.002(0.174) 0.965(0.177) 0.996(0.153) 1.013(0.178) 1.012(0.202)
#>   2  0.998(0.173) 0.966(0.182) 0.984(0.182) 1.008(0.171) 0.983(0.209)
#>   3  0.993(0.239) 0.985(0.206) 0.995(0.213) 1.013(0.209) 1.008(0.191)
#>   4  0.979(0.232) 0.998(0.185) 1.026(0.211) 0.966(0.217) 0.970(0.205)
#>   5  1.008(0.182) 0.997(0.236) 1.009(0.144) 1.021(0.184) 0.982(0.188)
#>   6  0.976(0.191) 1.004(0.181) 0.963(0.196) 0.961(0.187) 0.968(0.183)
#>   7  1.004(0.192) 1.019(0.208) 0.954(0.175) 0.992(0.197) 1.010(0.231)
#>   8  0.948(0.187) 0.960(0.224) 0.992(0.152) 0.988(0.221) 0.989(0.177)
#>   9  1.030(0.203) 0.975(0.162) 0.997(0.186) 0.984(0.184) 1.015(0.222)
#>   10 1.038(0.182) 0.995(0.185) 0.982(0.210) 1.003(0.220) 0.976(0.186)
#>     year
#> age  1975         1976         1977         1978         1979        
#>   1  0.958(0.216) 0.999(0.192) 0.971(0.230) 1.030(0.184) 0.984(0.203)
#>   2  0.991(0.213) 1.023(0.226) 0.996(0.235) 0.985(0.197) 1.018(0.176)
#>   3  1.029(0.207) 1.015(0.197) 1.017(0.189) 1.004(0.224) 0.998(0.195)
#>   4  1.006(0.205) 1.004(0.175) 1.013(0.222) 0.983(0.193) 0.985(0.210)
#>   5  1.056(0.218) 0.971(0.175) 0.973(0.244) 1.018(0.205) 1.042(0.188)
#>   6  0.992(0.200) 0.978(0.202) 1.050(0.219) 1.039(0.193) 0.970(0.201)
#>   7  0.967(0.172) 0.983(0.197) 0.989(0.203) 0.969(0.197) 1.030(0.200)
#>   8  0.997(0.160) 0.990(0.183) 0.970(0.191) 0.984(0.203) 1.013(0.242)
#>   9  0.981(0.193) 0.976(0.182) 0.981(0.208) 1.042(0.219) 1.020(0.176)
#>   10 0.999(0.181) 0.976(0.194) 0.965(0.185) 0.963(0.218) 1.003(0.189)
#>     year
#> age  1980         1981         1982         1983         1984        
#>   1  1.082(0.209) 0.994(0.217) 1.076(0.224) 0.997(0.217) 0.985(0.185)
#>   2  1.007(0.183) 0.997(0.200) 0.991(0.147) 1.041(0.240) 0.956(0.200)
#>   3  1.011(0.212) 1.021(0.162) 1.001(0.172) 1.001(0.200) 1.025(0.160)
#>   4  0.993(0.218) 0.987(0.168) 1.000(0.185) 0.969(0.221) 1.024(0.232)
#>   5  0.990(0.159) 1.012(0.222) 1.029(0.236) 0.982(0.183) 1.007(0.208)
#>   6  0.987(0.211) 1.001(0.205) 0.985(0.166) 0.987(0.211) 1.035(0.194)
#>   7  1.009(0.175) 0.999(0.170) 1.030(0.187) 1.004(0.193) 1.030(0.204)
#>   8  1.004(0.205) 0.987(0.164) 1.006(0.202) 1.008(0.189) 1.005(0.204)
#>   9  0.965(0.158) 1.034(0.257) 1.009(0.178) 0.971(0.179) 1.016(0.213)
#>   10 0.995(0.172) 1.056(0.237) 0.988(0.212) 0.987(0.194) 1.043(0.224)
#>     year
#> age  1985         1986         1987         1988         1989        
#>   1  0.984(0.180) 0.982(0.203) 0.992(0.182) 0.991(0.211) 1.000(0.169)
#>   2  0.979(0.196) 0.972(0.218) 0.998(0.222) 1.036(0.209) 1.017(0.218)
#>   3  1.002(0.225) 1.040(0.218) 1.027(0.172) 1.036(0.178) 1.022(0.187)
#>   4  0.966(0.213) 1.012(0.200) 0.998(0.165) 0.983(0.185) 0.998(0.215)
#>   5  0.963(0.180) 1.014(0.189) 0.983(0.167) 0.993(0.166) 0.999(0.215)
#>   6  1.017(0.194) 0.998(0.189) 0.945(0.175) 1.020(0.179) 1.040(0.238)
#>   7  1.033(0.214) 0.998(0.158) 0.973(0.205) 1.004(0.201) 1.022(0.182)
#>   8  0.976(0.165) 0.997(0.191) 1.000(0.207) 1.003(0.249) 1.011(0.198)
#>   9  1.005(0.228) 0.987(0.188) 1.007(0.240) 1.025(0.229) 0.982(0.220)
#>   10 1.034(0.193) 0.999(0.186) 0.999(0.186) 0.984(0.214) 0.999(0.195)
#>     year
#> age  1990         1991         1992         1993         1994        
#>   1  0.969(0.206) 0.992(0.203) 0.989(0.184) 1.006(0.210) 1.023(0.200)
#>   2  1.049(0.219) 0.997(0.171) 0.979(0.227) 1.027(0.196) 1.032(0.167)
#>   3  1.017(0.187) 1.029(0.240) 0.960(0.211) 0.966(0.178) 0.998(0.170)
#>   4  0.986(0.207) 1.045(0.176) 1.006(0.190) 1.005(0.183) 0.998(0.215)
#>   5  1.015(0.224) 1.017(0.182) 1.039(0.195) 1.014(0.213) 1.001(0.208)
#>   6  1.002(0.173) 1.002(0.156) 1.014(0.204) 0.976(0.212) 0.979(0.208)
#>   7  1.000(0.162) 0.998(0.179) 0.981(0.205) 1.010(0.214) 1.005(0.208)
#>   8  1.050(0.230) 1.034(0.185) 0.983(0.157) 1.043(0.204) 1.017(0.217)
#>   9  0.974(0.203) 1.021(0.175) 0.952(0.203) 1.025(0.232) 0.975(0.175)
#>   10 1.015(0.180) 0.995(0.188) 1.021(0.172) 0.982(0.207) 1.014(0.237)
#>     year
#> age  1995         1996         1997         1998         1999        
#>   1  0.986(0.252) 1.062(0.240) 1.007(0.176) 0.965(0.191) 1.037(0.184)
#>   2  1.028(0.217) 0.994(0.211) 0.985(0.234) 0.983(0.205) 0.998(0.175)
#>   3  0.993(0.190) 0.975(0.172) 0.974(0.213) 0.990(0.181) 1.053(0.210)
#>   4  1.014(0.182) 1.045(0.225) 1.005(0.175) 1.046(0.207) 0.977(0.193)
#>   5  1.029(0.223) 0.937(0.202) 0.998(0.184) 1.056(0.188) 1.028(0.176)
#>   6  1.002(0.204) 0.966(0.174) 0.990(0.161) 0.987(0.200) 0.978(0.175)
#>   7  0.991(0.174) 1.027(0.161) 0.962(0.197) 0.972(0.209) 1.005(0.181)
#>   8  0.985(0.177) 0.973(0.181) 0.997(0.188) 1.003(0.194) 1.008(0.211)
#>   9  0.999(0.208) 0.988(0.212) 0.963(0.215) 1.030(0.179) 0.986(0.185)
#>   10 0.982(0.181) 1.027(0.198) 1.051(0.191) 0.999(0.212) 1.014(0.167)
#>     year
#> age  2000         2001         2002         2003         2004        
#>   1  0.978(0.190) 1.014(0.176) 1.018(0.216) 0.976(0.204) 1.007(0.223)
#>   2  0.989(0.192) 1.046(0.157) 1.009(0.195) 0.969(0.143) 0.995(0.159)
#>   3  0.999(0.198) 0.995(0.195) 1.023(0.197) 0.993(0.213) 1.014(0.213)
#>   4  1.021(0.139) 0.987(0.147) 0.973(0.234) 1.044(0.238) 1.019(0.187)
#>   5  1.021(0.204) 0.953(0.172) 1.027(0.223) 1.019(0.218) 1.029(0.211)
#>   6  1.037(0.185) 0.984(0.233) 0.973(0.197) 1.031(0.235) 1.003(0.196)
#>   7  0.979(0.226) 1.011(0.165) 0.977(0.214) 1.037(0.237) 1.029(0.191)
#>   8  1.014(0.190) 0.956(0.214) 1.015(0.217) 0.986(0.172) 0.986(0.207)
#>   9  1.023(0.206) 0.987(0.162) 0.980(0.184) 1.028(0.235) 1.076(0.260)
#>   10 0.995(0.252) 0.964(0.192) 1.019(0.175) 0.993(0.208) 1.006(0.220)
#>     year
#> age  2005         2006         2007         2008         2009        
#>   1  0.987(0.224) 1.023(0.197) 0.964(0.164) 0.997(0.239) 1.021(0.187)
#>   2  1.036(0.211) 1.045(0.205) 0.983(0.185) 1.045(0.215) 1.022(0.236)
#>   3  1.046(0.192) 1.008(0.181) 1.007(0.191) 1.015(0.189) 0.991(0.203)
#>   4  1.001(0.217) 1.006(0.191) 1.022(0.223) 0.974(0.240) 1.053(0.234)
#>   5  0.974(0.148) 0.978(0.171) 0.991(0.191) 0.997(0.211) 0.990(0.225)
#>   6  1.010(0.205) 1.012(0.197) 1.048(0.183) 0.999(0.224) 1.011(0.181)
#>   7  0.993(0.201) 0.987(0.189) 1.034(0.165) 0.977(0.181) 1.001(0.244)
#>   8  1.000(0.187) 0.965(0.194) 1.000(0.200) 1.013(0.203) 1.007(0.187)
#>   9  0.994(0.198) 1.021(0.184) 0.983(0.171) 1.065(0.219) 1.041(0.162)
#>   10 0.948(0.193) 1.005(0.170) 1.003(0.213) 1.003(0.202) 1.032(0.202)
#>     year
#> age  2010         2011         2012         2013         2014        
#>   1  1.016(0.197) 0.957(0.194) 0.976(0.202) 1.014(0.201) 0.992(0.163)
#>   2  0.996(0.222) 0.991(0.183) 1.026(0.184) 1.028(0.212) 1.036(0.173)
#>   3  0.988(0.196) 1.012(0.179) 0.998(0.179) 0.998(0.179) 0.980(0.193)
#>   4  0.978(0.163) 0.983(0.207) 1.020(0.183) 0.987(0.185) 1.004(0.204)
#>   5  0.999(0.233) 1.047(0.224) 1.005(0.197) 0.975(0.193) 0.988(0.203)
#>   6  0.976(0.164) 1.020(0.167) 1.010(0.202) 1.014(0.223) 0.963(0.197)
#>   7  1.002(0.159) 1.000(0.192) 1.008(0.216) 0.995(0.221) 1.002(0.191)
#>   8  0.963(0.195) 1.036(0.186) 0.993(0.197) 0.973(0.230) 0.972(0.215)
#>   9  0.961(0.183) 1.002(0.200) 0.989(0.186) 1.028(0.200) 1.010(0.234)
#>   10 0.994(0.177) 0.971(0.166) 1.030(0.233) 0.989(0.167) 0.960(0.167)
#>     year
#> age  2015         2016         2017         2018         2019        
#>   1  1.031(0.196) 1.000(0.222) 0.999(0.215) 0.996(0.194) 0.971(0.184)
#>   2  0.999(0.181) 0.973(0.208) 1.019(0.205) 0.966(0.181) 1.019(0.205)
#>   3  0.972(0.207) 0.976(0.167) 1.026(0.225) 0.986(0.190) 0.990(0.162)
#>   4  0.984(0.189) 1.023(0.229) 1.043(0.226) 1.038(0.187) 1.026(0.183)
#>   5  0.942(0.237) 0.989(0.144) 1.014(0.218) 1.044(0.225) 1.017(0.171)
#>   6  0.996(0.244) 0.974(0.182) 1.021(0.192) 0.995(0.208) 0.963(0.192)
#>   7  1.013(0.186) 1.016(0.183) 0.979(0.189) 1.001(0.224) 1.004(0.234)
#>   8  1.018(0.200) 0.960(0.192) 1.009(0.173) 1.027(0.205) 0.999(0.157)
#>   9  1.046(0.225) 0.988(0.174) 1.006(0.215) 1.007(0.195) 1.049(0.184)
#>   10 0.987(0.241) 1.021(0.182) 0.987(0.174) 0.995(0.170) 0.975(0.190)
#>     year
#> age  2020         2021         2022         2023         2024        
#>   1  0.981(0.213) 1.026(0.215) 0.986(0.164) 0.981(0.178) 0.986(0.189)
#>   2  0.996(0.202) 1.019(0.240) 1.026(0.187) 0.976(0.179) 1.009(0.227)
#>   3  0.978(0.240) 1.022(0.206) 1.020(0.185) 0.961(0.193) 1.001(0.176)
#>   4  0.938(0.153) 1.039(0.212) 1.004(0.131) 1.012(0.188) 0.970(0.186)
#>   5  0.946(0.184) 1.006(0.245) 1.019(0.205) 0.978(0.190) 0.999(0.190)
#>   6  1.030(0.205) 1.032(0.229) 0.962(0.177) 1.026(0.202) 0.982(0.178)
#>   7  0.976(0.182) 0.983(0.200) 0.981(0.219) 1.012(0.194) 0.977(0.194)
#>   8  1.027(0.195) 0.960(0.184) 1.015(0.217) 0.970(0.191) 1.026(0.192)
#>   9  0.992(0.208) 0.974(0.190) 0.998(0.180) 0.995(0.191) 1.025(0.246)
#>   10 0.976(0.199) 1.042(0.211) 0.981(0.196) 0.996(0.156) 0.999(0.206)
#>     year
#> age  2025         2026         2027         2028         2029        
#>   1  1.010(0.204) 1.012(0.196) 1.002(0.190) 1.020(0.233) 1.001(0.204)
#>   2  1.044(0.183) 1.008(0.225) 1.009(0.175) 1.009(0.203) 1.019(0.202)
#>   3  0.990(0.190) 1.024(0.203) 0.979(0.155) 1.001(0.231) 0.973(0.198)
#>   4  0.996(0.214) 1.063(0.258) 0.990(0.208) 0.959(0.172) 1.023(0.195)
#>   5  1.023(0.219) 0.991(0.218) 0.976(0.220) 0.981(0.217) 0.965(0.207)
#>   6  1.049(0.226) 0.995(0.201) 0.961(0.206) 0.983(0.198) 1.033(0.227)
#>   7  0.984(0.197) 1.025(0.170) 0.973(0.191) 1.025(0.225) 1.026(0.189)
#>   8  0.996(0.190) 0.994(0.173) 0.999(0.244) 0.982(0.197) 0.998(0.196)
#>   9  0.997(0.178) 1.017(0.246) 0.981(0.180) 0.970(0.199) 1.025(0.239)
#>   10 1.006(0.206) 1.036(0.246) 0.947(0.185) 1.013(0.226) 1.024(0.203)
#>     year
#> age  2030         2031         2032         2033         2034        
#>   1  0.984(0.228) 0.988(0.183) 1.003(0.257) 1.021(0.206) 0.939(0.163)
#>   2  1.001(0.231) 0.952(0.176) 1.043(0.220) 1.004(0.200) 1.012(0.222)
#>   3  1.028(0.223) 0.972(0.200) 1.004(0.219) 0.980(0.188) 0.964(0.198)
#>   4  1.008(0.200) 0.992(0.188) 1.025(0.187) 0.921(0.208) 0.953(0.159)
#>   5  0.973(0.225) 0.996(0.255) 1.003(0.229) 0.999(0.170) 0.985(0.191)
#>   6  1.004(0.161) 1.030(0.210) 1.002(0.199) 0.964(0.155) 0.990(0.216)
#>   7  1.000(0.188) 1.016(0.196) 0.994(0.198) 1.004(0.189) 0.976(0.215)
#>   8  0.995(0.183) 0.988(0.183) 1.020(0.181) 0.995(0.204) 1.016(0.256)
#>   9  1.008(0.200) 1.004(0.160) 1.017(0.173) 1.053(0.201) 1.019(0.207)
#>   10 1.016(0.221) 1.010(0.197) 0.996(0.194) 0.982(0.190) 0.947(0.189)
#>     year
#> age  2035         2036         2037         2038         2039        
#>   1  0.976(0.179) 1.043(0.152) 1.033(0.190) 1.008(0.152) 1.000(0.189)
#>   2  0.942(0.137) 1.019(0.206) 0.984(0.183) 0.990(0.157) 0.988(0.172)
#>   3  0.989(0.181) 1.010(0.230) 0.977(0.207) 0.953(0.180) 0.995(0.213)
#>   4  0.985(0.205) 1.002(0.163) 0.984(0.190) 1.015(0.177) 1.007(0.173)
#>   5  1.033(0.222) 1.015(0.200) 1.013(0.194) 1.000(0.192) 0.984(0.213)
#>   6  0.994(0.246) 0.980(0.167) 0.986(0.178) 1.007(0.176) 0.990(0.179)
#>   7  0.994(0.193) 0.992(0.198) 1.054(0.221) 1.002(0.185) 1.004(0.194)
#>   8  1.039(0.197) 0.959(0.208) 0.971(0.187) 0.995(0.171) 1.021(0.188)
#>   9  1.022(0.202) 0.965(0.198) 0.977(0.222) 0.972(0.195) 0.977(0.193)
#>   10 0.987(0.173) 1.041(0.210) 0.997(0.180) 0.985(0.173) 0.980(0.173)
#>     year
#> age  2040         2041         2042         2043         2044        
#>   1  0.985(0.196) 0.957(0.197) 0.995(0.213) 1.016(0.176) 1.003(0.203)
#>   2  1.000(0.260) 1.011(0.175) 0.980(0.174) 1.011(0.220) 1.008(0.172)
#>   3  1.034(0.216) 0.986(0.171) 1.026(0.194) 1.038(0.199) 1.046(0.189)
#>   4  1.029(0.228) 1.003(0.177) 1.011(0.214) 1.022(0.193) 1.008(0.214)
#>   5  1.011(0.201) 1.006(0.186) 1.026(0.212) 0.993(0.191) 1.007(0.126)
#>   6  1.016(0.194) 1.020(0.232) 0.976(0.203) 0.988(0.221) 0.993(0.188)
#>   7  1.018(0.232) 1.016(0.234) 1.005(0.165) 0.965(0.184) 0.957(0.251)
#>   8  0.976(0.239) 0.997(0.173) 1.024(0.216) 0.960(0.158) 0.990(0.170)
#>   9  0.963(0.170) 0.972(0.225) 1.043(0.203) 1.019(0.156) 0.979(0.220)
#>   10 1.063(0.200) 1.002(0.169) 1.013(0.228) 1.001(0.219) 1.040(0.180)
#>     year
#> age  2045         2046         2047         2048         2049        
#>   1  0.992(0.181) 1.050(0.186) 1.001(0.206) 0.946(0.189) 1.052(0.203)
#>   2  1.016(0.236) 0.985(0.175) 1.011(0.246) 0.973(0.170) 1.017(0.167)
#>   3  0.981(0.193) 0.982(0.187) 1.024(0.175) 1.019(0.223) 1.015(0.170)
#>   4  1.011(0.199) 1.014(0.228) 1.053(0.201) 1.010(0.210) 1.005(0.191)
#>   5  1.003(0.199) 1.010(0.216) 1.009(0.166) 0.976(0.193) 0.993(0.214)
#>   6  1.003(0.212) 1.007(0.193) 0.971(0.181) 1.022(0.205) 0.969(0.191)
#>   7  1.009(0.171) 1.006(0.178) 1.009(0.182) 0.994(0.194) 1.018(0.170)
#>   8  0.991(0.222) 1.012(0.215) 1.015(0.196) 1.018(0.206) 0.985(0.217)
#>   9  1.016(0.214) 0.987(0.196) 1.003(0.185) 1.009(0.195) 0.981(0.195)
#>   10 1.022(0.254) 1.058(0.200) 1.009(0.238) 1.012(0.196) 1.000(0.197)
#>     year
#> age  2050         2051         2052         2053         2054        
#>   1  0.983(0.195) 1.001(0.222) 1.010(0.205) 1.016(0.234) 1.041(0.233)
#>   2  1.044(0.237) 1.021(0.217) 0.985(0.214) 0.997(0.194) 0.956(0.198)
#>   3  0.984(0.160) 1.007(0.217) 1.067(0.194) 1.033(0.254) 1.008(0.238)
#>   4  0.968(0.183) 0.994(0.183) 1.007(0.200) 0.968(0.233) 0.996(0.194)
#>   5  0.957(0.202) 1.011(0.179) 1.001(0.183) 0.983(0.194) 0.989(0.211)
#>   6  0.980(0.176) 1.001(0.218) 1.007(0.173) 0.979(0.166) 0.953(0.161)
#>   7  1.057(0.214) 1.003(0.206) 0.967(0.176) 0.988(0.185) 1.020(0.204)
#>   8  0.997(0.152) 0.999(0.219) 1.043(0.229) 0.977(0.226) 1.014(0.194)
#>   9  1.000(0.203) 0.980(0.178) 0.965(0.162) 0.986(0.166) 0.999(0.191)
#>   10 0.976(0.190) 1.016(0.187) 1.026(0.196) 0.963(0.182) 0.979(0.201)
#>     year
#> age  2055        
#>   1  1.018(0.161)
#>   2  0.994(0.163)
#>   3  0.945(0.192)
#>   4  0.944(0.172)
#>   5  1.030(0.180)
#>   6  0.994(0.174)
#>   7  1.022(0.212)
#>   8  0.990(0.212)
#>   9  1.000(0.197)
#>   10 1.011(0.188)
#> 
#> units:  1000 
data(plesim)
x <- iter(oem, 1:50)
dims(observations(x)$stk)$iter
#> [1] 50
dims(deviances(x)$stk$catch.n)$iter
#> [1] 50
data(plesim)
iter(oem, 1:50)
#> Observation Error Model
#> Observations of: stk idx 
#> 
#> Deviances for stk:
#> An object of class "FLQuants"
#> 
#> Elements: catch.n 
#> 
#> Name: catch.n 
#> 	dim  :  10 96 1 1 1 50 
#> 	quant:  age 
#> 	units:  1000 
#> 
#> 	Min    :  0.4482633 
#> 	1st Qu.:  0.8752356 
#> 	Mean   :  1.020474 
#> 	Median :  1.000279 
#> 	3rd Qu.:  1.144219 
#> 	Max    :  2.204301 
#> 	NAs    :  0 %
#>  
#> 
#> Deviances for idx:
#> An object of class "FLQuants"
#> 
#> Elements: SUR CPUE 
#> 
#> Name: SUR 
#> 	dim  :  1 76 1 1 1 50 
#> 	quant:  quant 
#> 	units:  NA 
#> 
#> 	Min    :  0.4507767 
#> 	1st Qu.:  0.8714447 
#> 	Mean   :  1.022074 
#> 	Median :  1.003233 
#> 	3rd Qu.:  1.151817 
#> 	Max    :  2.457981 
#> 	NAs    :  0 %
#> Name: CPUE 
#> 	dim  :  1 76 1 1 1 50 
#> 	quant:  quant 
#> 	units:  NA 
#> 
#> 	Min    :  0.3655982 
#> 	1st Qu.:  0.8192952 
#> 	Mean   :  1.045867 
#> 	Median :  0.998022 
#> 	3rd Qu.:  1.225497 
#> 	Max    :  2.895648 
#> 	NAs    :  0 %
#>