A method to project the operating model (OM)

Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque eleifend odio ac rutrum luctus. Aenean placerat porttitor commodo. Pellentesque eget porta libero. Pellentesque molestie mi sed orci feugiat, non mollis enim tristique. Suspendisse eu sapien vitae arcu lobortis ultrices vitae ac velit. Curabitur id

fwd.om(stk, ctrl, sr, ...)

Arguments

ctrl
the fwdControl object with objectives and constraints

sr
a FLSR with the stock-recruitment model

object
the OM as a FLStock

sr.residuals
a FLQuant with S/R residuals

sr.residuals.mult
logical about residuals being multiplicative