The typical HCR used by ICES

The typical HCR used by ICES which sets a target F based on the SSB based on 4 parameters: sblim, sbsafe, fmin and ftrg. F increases linearly between SSB = blim and SSB = bsafe, from F = fmin to F = ftrg. If:

B < Blim, F = Fbycatch;

B > trigger, F = Fmsy;

B > Blim & B < trigger, F linear between Fbycatch and Fmsy;

F = ftrg is the maximum F, F = fmin is the minimum F. F is set in year ay, based on SSB in year ay - data_lag

ices.hcr(
  stk,
  ftrg,
  sblim,
  sbsafe,
  fmin = 0,
  minfbar = range(stk, "minfbar"),
  maxfbar = range(stk, "maxfbar"),
  args,
  tracking
)

Arguments

stk

The perceived FLStock.

ftrg

TODO:description

sblim

TODO:description

sbsafe

TODO:description

fmin

Minimum fishing mortality.

args

MSE arguments, class list.

tracking

Structure for tracking modules outputs.

Value

A list with elements ctrl, of class fwdControl, and tracking.

Examples

data(ple4)
# Test for year when SSB > bsafe
ices.hcr(ple4, fmin=0.05, ftrg=0.15, sblim=200000, sbsafe=300000,
  args=list(ay=2018, data_lag=1, management_lag=1), tracking=FLQuant())
#> $ctrl
#> An object of class "fwdControl"
#>  (step) year quant min value max
#>       1 2019  fbar  NA 0.150  NA
#> 
#> $tracking
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1 
#>   all NA
#> 
#> units:  NA 
#> 
# Test for year when SSB < bsafe
ices.hcr(ple4, fmin=0.05, ftrg=0.15, sblim=200000, sbsafe=300000,
  args=list(ay=1995, data_lag=1, management_lag=1), tracking=FLQuant())
#> $ctrl
#> An object of class "fwdControl"
#>  (step) year quant min value max
#>       1 1996  fbar  NA 0.083  NA
#> 
#> $tracking
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1 
#>   all NA
#> 
#> units:  NA 
#>