mp executes a single run of a Management Procedure

mp executes a single run of a Management Procedure

mp(
  om,
  oem = NULL,
  iem = NULL,
  ctrl,
  args,
  scenario = "NA",
  tracking = "missing",
  verbose = TRUE,
  parallel = TRUE
)

Arguments

om
The operating model (OM), an object of class FLom or FLombf.

oem
The observation error model (OEM), an object of class FLoem.

iem
The implementation error model (IEM), an object of class FLiem.

ctrl
A control structure for the MP run, an object of class mpCtrl.

args
MSE arguments, list. Only 'iy', the starting year, is required.

scenario
Name of the scenario tested in this run, character.

tracking
Extra elements to add to the standard tracking FLQuant in its first dimensions, character.

verbose
Should output be verbose or not, logical.

Value

An object of class FLmse.

Examples

# [TODO:example]