S4 class mpCtrl

The mpCtrl class defines which modules will be run my a call to the mp function. It contains a series of objects of class mseCtrl only for those modules required by the defined MP.

# S4 method for mpCtrl
initialize(.Object, ...)

est(object, ...)

# S4 method for mpCtrl
est(object)

est(object) <- value

# S4 method for mpCtrl,`function`
est(object) <- value

phcr(object, ...)

# S4 method for mpCtrl
phcr(object)

phcr(object) <- value

# S4 method for mpCtrl,`function`
phcr(object) <- value

hcr(object, ...)

# S4 method for mpCtrl
hcr(object)

hcr(object) <- value

# S4 method for mpCtrl,`function`
hcr(object) <- value

isys(object, ...)

# S4 method for mpCtrl
isys(object)

isys(object) <- value

# S4 method for mpCtrl,`function`
isys(object) <- value

tm(object, ...)

# S4 method for mpCtrl
tm(object)

tm(object) <- value

# S4 method for mpCtrl,`function`
tm(object) <- value

# S4 method for mpCtrl
show(object)

# S4 method for mpCtrl
iters(object, iter)

# S4 method for mpCtrl
method(object, element)

# S4 method for mpCtrl,`function`
method(object, element) <- value

# S4 method for mpCtrl,`function`
args(object, element) <- value

# S4 method for mpCtrl,character
debug(fun, text)

Arguments

...

additional argument list that might never be used

object

object of relevant class (see signature of method)

Slots

est

Specification for the stock status estimator, class mseCtrl.

phcr

Specification for the harvest control rule parametrization, class mseCtrl.

hcr

Specification for the harvest control rule, class mseCtrl.

isys

Specification for the implementation system, class mseCtrl.

tm

Specification for technical measures, class mseCtrl.

Accessors

All slots in the class have accessor and replacement methods defined that allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot. A numeric vector can also be used when replacing FLQuant slots, and the vector will be used to substitute the values in the slot, but not its other attributes.

Constructor

A construction method exists for this class that can take named arguments for any of its slots. All slots are then created to match the requirements of the class validity. If an unnamed FLQuant object is provided, this is used for sizing, but not for populating any slot.