TAC implementation function

Performs a short term forecast (STF) to hit the target F in year ay+1. The resulting catch in year ay+1 is the TAC, i.e. the TAC that will result in Fbar = Ftarget. The STF uses geometric mean recruitment. Fbar in the intermediate years (i.e. years between last data year and ay+1) are set as the mean of the last nsqy years. The control argument is a list of parameters: nsqy - number of years to average over to get Fbar for STF delta_tac_min - constraint on the TAC delta_tac_max - constraint on the TAC

tac.is(stk, ctrl, ay, nsqy = 3, delta_tac_max = NA,
  delta_tac_min = NA, tracking)

Arguments

stk
The perceived FLStock.

ay
The year for which the target F is set, based on the SSB in year (ay - control$ssb_lag).

imp_control
A list with the elements: nsqy, delta_tac_min, delta_tac_max

EFF0
The tracking array

EFF
Not used by this function but may be used by the other implementation functions