Target-based harvest control rule to adjust input or output from CPUE

Short description

target.hcr(
  ind,
  lim,
  target,
  r = 1,
  metric = "mlc",
  output = "catch",
  nyears = 3,
  args,
  tracking
)

Value

A list containing ctrl, a fwdControl object, and tracking, an FLQuant.

Details

$$$$

References

Hoshino, E., Hillary, R., Davies, C., Satria, F., Sadiyah, L., Ernawati, T., and Proctor, C. 2020. Development of pilot empirical harvest strategies for tropical tuna in indonesian archipelagic waters: case studies of skipjack and yellowfin tuna. Fisheries Research, 227:105539, doi:10.1016/j.fishres.2020.105539.

See also

Author

The FLR Team

Examples

data(sol274)
#> Warning: namespace ‘AAP’ is not available and has been replaced
#> by .GlobalEnv when processing object ‘om’
#
est <- cpue.ind(stock(om), FLIndices(CPUE=FLIndexBiomass(index=ssb(om))),
  args=list(ay=2000, data_lag=1),
  tracking=FLQuant(dimnames=list(metric="ind", year=2000, iter=1:100)))
#
target.hcr(ind=est$ind, lim=28000, target=40000,
  metric="wmean", output="catch",
  args=list(ay=2000, frq=1, data_lag=1, management_lag=1),
  tracking=FLQuants(sol174=FLQuant(1000, dimnames=list(metric="hcr",
  year=2000))))
#> $ctrl
#> An object of class "fwdControl"
#>  (step) year quant fishery catch min        value max
#>       1 2001 catch       1     1  NA 0.912(0.085)  NA
#>    iters:  100 
#> 
#> 
#> $tracking
#> $ sol174 
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>       year
#> metric 2000
#>    hcr 1000
#> 
#> units:  NA 
#> 
#>