S4 class FLmse

The FLmse class stores information relative to the MSE's management procedure'.

FLmse(...)

FLmse(...)

om(object, ...)

# S4 method for FLmse
om(object)

om(object) <- value

tracking(object, ...)

# S4 method for FLmse
tracking(object, biol = "missing")

tracking(object, ...) <- value

control(object, ...)

# S4 method for FLmse
control(object, i = "missing")

control(object, ...) <- value

oem(object, ...)

# S4 method for FLmse
oem(object)

oem(object) <- value

# S4 method for FLmse
args(name)

# S4 method for FLmse,list
args(object) <- value

Arguments

...
additional argument list that might never be used

object
object of relevant class (see signature of method)

value
the new object

Slots

om

FLom with the operating model.

tracking

FLQuant with record of decisions made during the mp cycle.

args

list with assorted arguments required to run the MSE cycle.

Accessors

All slots in the class have accessor and replacement methods defined that allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot. A numeric vector can also be used when replacing FLQuant slots, and the vector will be used to substitute the values in the slot, but not its other attributes.

Constructor

A construction method exists for this class that can take named arguments for any of its slots. All slots are then created to match the requirements of the class validity. If an unnamed FLQuant object is provided, this is used for sizing, but not for populating any slot.

Examples