Specification for the observation error model (OEM).

The FLoem class stores the method, arguments and observations that define the way observations are collected from an operating model at each time step in the management procedure. This class extends mseCtrl through the addition of two lists used to gather past and new observations, and deviances to use on each step in the observation process.

FLoem(...)

FLoem(...)

observations(object, ...)

# S4 method for FLoem
observations(object, ...)

observations(object, i) <- value

# S4 method for FLoem,missing,list
observations(object) <- value

# S4 method for FLoem,ANY,FLStock
observations(object, i) <- value

deviances(object, ...)

# S4 method for FLoem
deviances(object)

deviances(object) <- value

# S4 method for FLoem,list
deviances(object) <- value

# S4 method for FLoem
show(object)

# S4 method for FLoem
iter(obj, iter)

# S4 method for FLoem,FLoem
combine(x, y, ..., check = FALSE)

Arguments

...

additional argument list that might never be used

object

object of relevant class (see signature of method)

value

the new object

Slots

method

The function to be run in the module call, class function.

args

Arguments to be passed to the method, of class list.

observations

Past observations, class list.

deviances

Observation deviances, class list.

Accessors

All slots in the class have accessor and replacement methods defined that allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot. A numeric vector can also be used when replacing FLQuant slots, and the vector will be used to substitute the values in the slot, but not its other attributes.

Constructor

A construction method exists for this class that can take named arguments for any of its slots. All slots are then created to match the requirements of the class validity. If an unnamed FLQuant object is provided, this is used for sizing, but not for populating any slot.