Initializes a population for a given virgin biomass.

Abundances at age for a population at virgin conditions at age. An FLBiol object is initiated by providing a target total biomass (B0) and a value for the stock-recruit steepness (s). The object requires slots to be already filled up for the mean weight-at-age (wt), natural mortality (m), time of spawning (spwn) and maturity at age (mat).

initiate(biol, B0, h = 0.75)

Arguments

biol

An FLBiol object to nbe initiated.

B0

Initial or virgin biomass.

Value

An updated FLBiol with abundances set in teh first year to match the requested biomas.

Examples

data(ple4.biol)
initiate(ple4.biol, B0=450000)
#> [1] 1
#> An object of class "FLBiol"
#> 
#> Name: PLE 
#> Description: Plaice in IV. ICES WGNSSK 2018. FLAAP 
#> Quant: age 
#> Dims:  age 	year	unit	season	area	iter
#> 	10	61	1	1	1	1	
#> 
#> Range:  min	max	pgroup	minyear	maxyear 
#> 	1	10	10	1957	2017	
#> 
#> mat           ~ mat 
#>   mat         : [ 10 61 1 1 1 1 ], units =   
#>   NA          : [ 1 1 ], units =  NA 
#> fec           ~ fec 
#>   fec         : [ 10 61 1 1 1 1 ], units =   
#>   NA          : [ 1 1 ], units =  NA 
#> rec           ~ rec (4 * s * R0 * ssb)/(v * (1 - s) + ssb * (5 * s - 1)) 
#>   s, R0, v    : [ 3 1 ], units =  NA 
#
initiate(propagate(ple4.biol, 100), B0=runif(100, 3e5, 5e5))
#> [1] 1
#> [1] 2
#> [1] 3
#> [1] 4
#> [1] 5
#> [1] 6
#> [1] 7
#> [1] 8
#> [1] 9
#> [1] 10
#> [1] 11
#> [1] 12
#> [1] 13
#> [1] 14
#> [1] 15
#> [1] 16
#> [1] 17
#> [1] 18
#> [1] 19
#> [1] 20
#> [1] 21
#> [1] 22
#> [1] 23
#> [1] 24
#> [1] 25
#> [1] 26
#> [1] 27
#> [1] 28
#> [1] 29
#> [1] 30
#> [1] 31
#> [1] 32
#> [1] 33
#> [1] 34
#> [1] 35
#> [1] 36
#> [1] 37
#> [1] 38
#> [1] 39
#> [1] 40
#> [1] 41
#> [1] 42
#> [1] 43
#> [1] 44
#> [1] 45
#> [1] 46
#> [1] 47
#> [1] 48
#> [1] 49
#> [1] 50
#> [1] 51
#> [1] 52
#> [1] 53
#> [1] 54
#> [1] 55
#> [1] 56
#> [1] 57
#> [1] 58
#> [1] 59
#> [1] 60
#> [1] 61
#> [1] 62
#> [1] 63
#> [1] 64
#> [1] 65
#> [1] 66
#> [1] 67
#> [1] 68
#> [1] 69
#> [1] 70
#> [1] 71
#> [1] 72
#> [1] 73
#> [1] 74
#> [1] 75
#> [1] 76
#> [1] 77
#> [1] 78
#> [1] 79
#> [1] 80
#> [1] 81
#> [1] 82
#> [1] 83
#> [1] 84
#> [1] 85
#> [1] 86
#> [1] 87
#> [1] 88
#> [1] 89
#> [1] 90
#> [1] 91
#> [1] 92
#> [1] 93
#> [1] 94
#> [1] 95
#> [1] 96
#> [1] 97
#> [1] 98
#> [1] 99
#> [1] 100
#> An object of class "FLBiol"
#> 
#> Name: PLE 
#> Description: Plaice in IV. ICES WGNSSK 2018. FLAAP 
#> Quant: age 
#> Dims:  age 	year	unit	season	area	iter
#> 	10	61	1	1	1	100	
#> 
#> Range:  min	max	pgroup	minyear	maxyear 
#> 	1	10	10	1957	2017	
#> 
#> mat           ~ mat 
#>   mat         : [ 10 61 1 1 1 100 ], units =   
#>   NA          : [ 1 100 ], units =  NA 
#> fec           ~ fec 
#>   fec         : [ 10 61 1 1 1 100 ], units =   
#>   NA          : [ 1 100 ], units =  NA 
#> rec           ~ rec (4 * s * R0 * ssb)/(v * (1 - s) + ssb * (5 * s - 1)) 
#>   s, R0, v    : [ 3 100 ], units =  NA