A perfect observation of catch and abundances-at-age.

This observation error model function generates a set of perfect observations on catches, biology and abundance. Direct observations are made of the stock, while a single age-structured index of abundance, in numbers, is created with a fixed catchability of 0.01. deviances on either stk$catch.n or idx$index, if given, are applied.

perfect.oem(om, deviances, observations, args, tracking, biomass = FALSE, ...)

Arguments

om
An operating model, class FLom or FLombf.

deviances
A named list of observation deviances, class list.

observations
A list of past observations, extended to the end of om, class list.

args
The mp dimensions arguments, as generated by mp, class list.

tracking
Object to track module decisions and outputs, class FLQuant.

Value

A named list with elements stk (FLStock), idx (FLIndices), deviances, observations and tracking.

Details

This oem function generates a full observation time series every time step, and does not append them to existing objects in observations.

Examples

# On FLom
data(ple4om)
obs <- perfect.oem(om, deviances=NULL, observations=NULL,
  args=list(y0=1957, dy=2017), tracking=FLQuant())