A perfect 'estimate' of abundances, catches and harvest.

The FLStock generated by the call to oem is simply passed on in this function. The estimates of abundance, catches and exploitation will thus be as precise as the OEM observation.

perfect.sa(stk, idx, args, tracking, ...)

Arguments

stk

The stock observation generated by oem. Class FLStock.

idx

An observation of chnages in abundance, not used. Class FLIndices.

args

MSE arguments, class list.

tracking

Structure for tracking modules outputs.

Value

A list with elements stk and tracking.

Examples

data(sol274)
#> Warning: namespace ‘AAP’ is not available and has been replaced
#> by .GlobalEnv when processing object ‘om’
perfect.sa(stock(om), FLIndices(), args=list(ay=2018, dy=2017),
  tracking=FLQuants(FLQuant(dimnames=list(metric="conv.est", year=2018))))
#> $stk
#> An object of class "FLStock"
#> 
#> Name: Sole in North Sea 
#> Description: Imported from a VPA file. ( bootstrap/data/index.txt ).  Fri Apr 30 15:37: [...] 
#> Quant: age 
#> Dims:  age 	year	unit	season	area	iter
#> 	10	61	1	1	1	100	
#> 
#> Range:  min	max	pgroup	minyear	maxyear	minfbar	maxfbar 
#> 	1	10	10	1957	2017	2	6	
#> 
#> Metrics: 
#>   rec: 10507 - 644880  (thousands) 
#>   ssb: 17873 - 119187  (tonnes) 
#>   catch: 11339 - 35117  (tonnes) 
#>   fbar: 0.21 - 0.70  (f) 
#> 
#> $tracking
#> $ NA 
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>           year
#> metric     2018
#>   conv.est 1   
#> 
#> units:  NA 
#> 
#>