Compute performance statistics

TODO

# S4 method for FLQuants
performance(
  x,
  statistics,
  refpts = FLPar(),
  years = setNames(list(dimnames(x[[1]])$year), nm = dims(x[[1]])$maxyear),
  probs = c(0.1, 0.25, 0.5, 0.75, 0.9),
  mp = NULL
)

# S4 method for FLStock
performance(
  x,
  statistics,
  refpts = FLPar(),
  years = as.character(seq(dims(x)$minyear, dims(x)$maxyear)),
  metrics = FLCore::metrics(x),
  probs = NULL,
  mp = NULL
)

# S4 method for FLStocks
performance(
  x,
  statistics,
  refpts = FLPar(),
  years = dims(x[[1]])$maxyear,
  metrics = FLCore::metrics,
  probs = NULL,
  grid = missing,
  mp = NULL,
  mc.cores = 1
)

# S4 method for list
performance(
  x,
  statistics,
  refpts = FLPar(),
  years = dims(x[[1]])$maxyear,
  probs = NULL,
  grid = "missing",
  mp = NULL,
  mc.cores = 1,
  ...
)

# S4 method for FLom
performance(x, refpts = x@refpts, ...)

Arguments

statistics
statistics to be computed, as formula, name and description, list

refpts
Reference points for calculations, list

years
Years on which statistics should be computed, defaults to last year of input FLQuants

run
Object holding the results of forward projections, as a named FLQuants

Value

data.table Results of computing performance statistics.

Details

Each statistics is an object of class list object, with three elements, the first two of them compulsory:

An unnamed element of class formula, e.g. yearMeans(SB/SB0).

name: A short name to be output on tables and plots, of class character, e.g. "SB/SB0".

desc: A longer description of the statistics, of class character, e.g. "Mean spawner biomass relative to unfished"

Each statistic formula is evaluated against the metrics and refpts used in the function call. Formulas can thus use (i) the names of the FLQuants object or of the object returned by the call to metrics(), (ii) of the params in the refpts object and, for all classes but FLQuants, (iii) functions that can be called on object. See examples below for the necessary matching between metrics, refpts and the statistics formulas.

See also

Author

Iago Mosqueira, EC JRC

Examples

# LOAD example FLmse object
data(ple4om)
# GENERATE run from last 16 years of OM
run <- window(stock(om), start=2000, end=2015)
# DEFINE statistics
statistics <- list(
  dCatch=list(~yearMeans(C[, -1]/C[, -dims(C)$year]),
    name="mean(C[t] / C[t-1])",
    desc="Mean absolute proportional change in catch"),
  varCatch=list(~yearVars(C),
    name="var(C)",
    desc="Variance in catch"),
  varF=list(~yearVars(F),
    name="var(F)",
    desc="Variance in fishing mortality"))
# COMPUTE performance
performance(run, statistics, refpts=FLPar(MSY=110000),
  metrics=list(C=catch, F=fbar), years=list(2000:2015))

#>     statistic year         data iter                name
#>  1:    dCatch 2015 9.994373e-01    1 mean(C[t] / C[t-1])
#>  2:    dCatch 2015 9.994373e-01    2 mean(C[t] / C[t-1])
#>  3:    dCatch 2015 9.984355e-01    3 mean(C[t] / C[t-1])
#>  4:    dCatch 2015 9.986469e-01    4 mean(C[t] / C[t-1])
#>  5:    dCatch 2015 9.994853e-01    5 mean(C[t] / C[t-1])
#>  6:    dCatch 2015 9.989960e-01    6 mean(C[t] / C[t-1])
#>  7:    dCatch 2015 9.986956e-01    7 mean(C[t] / C[t-1])
#>  8:    dCatch 2015 9.985410e-01    8 mean(C[t] / C[t-1])
#>  9:    dCatch 2015 9.988764e-01    9 mean(C[t] / C[t-1])
#> 10:    dCatch 2015 9.984006e-01   10 mean(C[t] / C[t-1])
#> 11:    dCatch 2015 9.972845e-01   11 mean(C[t] / C[t-1])
#> 12:    dCatch 2015 9.980538e-01   12 mean(C[t] / C[t-1])
#> 13:    dCatch 2015 9.987969e-01   13 mean(C[t] / C[t-1])
#> 14:    dCatch 2015 9.992097e-01   14 mean(C[t] / C[t-1])
#> 15:    dCatch 2015 9.997025e-01   15 mean(C[t] / C[t-1])
#> 16:    dCatch 2015 1.000126e+00   16 mean(C[t] / C[t-1])
#> 17:    dCatch 2015 1.000371e+00   17 mean(C[t] / C[t-1])
#> 18:    dCatch 2015 1.000372e+00   18 mean(C[t] / C[t-1])
#> 19:    dCatch 2015 1.000083e+00   19 mean(C[t] / C[t-1])
#> 20:    dCatch 2015 9.996324e-01   20 mean(C[t] / C[t-1])
#> 21:    dCatch 2015 9.991367e-01   21 mean(C[t] / C[t-1])
#> 22:    dCatch 2015 9.997361e-01   22 mean(C[t] / C[t-1])
#> 23:    dCatch 2015 1.000617e+00   23 mean(C[t] / C[t-1])
#> 24:    dCatch 2015 9.991421e-01   24 mean(C[t] / C[t-1])
#> 25:    dCatch 2015 9.996747e-01   25 mean(C[t] / C[t-1])
#> 26:  varCatch 2015 2.291724e+08    1              var(C)
#> 27:  varCatch 2015 2.291724e+08    2              var(C)
#> 28:  varCatch 2015 2.225521e+08    3              var(C)
#> 29:  varCatch 2015 2.346856e+08    4              var(C)
#> 30:  varCatch 2015 2.349722e+08    5              var(C)
#> 31:  varCatch 2015 2.355825e+08    6              var(C)
#> 32:  varCatch 2015 2.266701e+08    7              var(C)
#> 33:  varCatch 2015 2.109882e+08    8              var(C)
#> 34:  varCatch 2015 2.268825e+08    9              var(C)
#> 35:  varCatch 2015 2.213593e+08   10              var(C)
#> 36:  varCatch 2015 2.118339e+08   11              var(C)
#> 37:  varCatch 2015 2.146798e+08   12              var(C)
#> 38:  varCatch 2015 2.151345e+08   13              var(C)
#> 39:  varCatch 2015 2.303325e+08   14              var(C)
#> 40:  varCatch 2015 2.232798e+08   15              var(C)
#> 41:  varCatch 2015 2.273611e+08   16              var(C)
#> 42:  varCatch 2015 2.182051e+08   17              var(C)
#> 43:  varCatch 2015 2.207349e+08   18              var(C)
#> 44:  varCatch 2015 2.330167e+08   19              var(C)
#> 45:  varCatch 2015 2.382070e+08   20              var(C)
#> 46:  varCatch 2015 2.320905e+08   21              var(C)
#> 47:  varCatch 2015 2.347540e+08   22              var(C)
#> 48:  varCatch 2015 2.334668e+08   23              var(C)
#> 49:  varCatch 2015 2.240261e+08   24              var(C)
#> 50:  varCatch 2015 2.299363e+08   25              var(C)
#> 51:      varF 2015 2.180638e-02    1              var(F)
#> 52:      varF 2015 2.180638e-02    2              var(F)
#> 53:      varF 2015 2.109697e-02    3              var(F)
#> 54:      varF 2015 2.216665e-02    4              var(F)
#> 55:      varF 2015 2.023697e-02    5              var(F)
#> 56:      varF 2015 1.989139e-02    6              var(F)
#> 57:      varF 2015 2.085215e-02    7              var(F)
#> 58:      varF 2015 1.919674e-02    8              var(F)
#> 59:      varF 2015 1.836105e-02    9              var(F)
#> 60:      varF 2015 1.963375e-02   10              var(F)
#> 61:      varF 2015 2.004514e-02   11              var(F)
#> 62:      varF 2015 2.018918e-02   12              var(F)
#> 63:      varF 2015 2.014899e-02   13              var(F)
#> 64:      varF 2015 1.940458e-02   14              var(F)
#> 65:      varF 2015 1.900968e-02   15              var(F)
#> 66:      varF 2015 2.096307e-02   16              var(F)
#> 67:      varF 2015 2.130031e-02   17              var(F)
#> 68:      varF 2015 2.142248e-02   18              var(F)
#> 69:      varF 2015 2.166655e-02   19              var(F)
#> 70:      varF 2015 2.143131e-02   20              var(F)
#> 71:      varF 2015 2.082809e-02   21              var(F)
#> 72:      varF 2015 2.129481e-02   22              var(F)
#> 73:      varF 2015 1.948213e-02   23              var(F)
#> 74:      varF 2015 2.126185e-02   24              var(F)
#> 75:      varF 2015 2.097983e-02   25              var(F)
#>     statistic year         data iter                name

# Minimum statistic, named list with formula and name
performance(run, statistics=list(CMSY=list(~C/MSY, name="CMSY")),
  refpts=FLPar(MSY=110000), metrics=list(C=catch, F=fbar),
  years=list(2000:2015))

#>      statistic year     data iter name
#>   1:      CMSY 2015 1.339249    1 CMSY
#>   2:      CMSY 2015 1.150833    1 CMSY
#>   3:      CMSY 2015 1.297908    1 CMSY
#>   4:      CMSY 2015 1.412723    1 CMSY
#>   5:      CMSY 2015 1.273223    1 CMSY
#>  ---                                  
#> 396:      CMSY 2015 1.086449   25 CMSY
#> 397:      CMSY 2015 1.216778   25 CMSY
#> 398:      CMSY 2015 1.289115   25 CMSY
#> 399:      CMSY 2015 1.261346   25 CMSY
#> 400:      CMSY 2015 1.250653   25 CMSY

# return quantiles
performance(run, statistics, refpts=FLPar(MSY=110000),
  metrics=list(C=catch, F=fbar), years=list(2000:2015),
  probs =  c(0.05, 0.1, 0.25, 0.5, 0.75, 0.9, 0.95))

#>     statistic year                name         data prob
#>  1:    dCatch 2015 mean(C[t] / C[t-1]) 9.981231e-01 0.05
#>  2:    dCatch 2015 mean(C[t] / C[t-1]) 9.984146e-01 0.10
#>  3:    dCatch 2015 mean(C[t] / C[t-1]) 9.986956e-01 0.25
#>  4:    dCatch 2015 mean(C[t] / C[t-1]) 9.992097e-01 0.50
#>  5:    dCatch 2015 mean(C[t] / C[t-1]) 9.997025e-01 0.75
#>  6:    dCatch 2015 mean(C[t] / C[t-1]) 1.000273e+00 0.90
#>  7:    dCatch 2015 mean(C[t] / C[t-1]) 1.000372e+00 0.95
#>  8:  varCatch 2015              var(C) 2.124031e+08 0.05
#>  9:  varCatch 2015              var(C) 2.148617e+08 0.10
#> 10:  varCatch 2015              var(C) 2.213593e+08 0.25
#> 11:  varCatch 2015              var(C) 2.273611e+08 0.50
#> 12:  varCatch 2015              var(C) 2.330167e+08 0.75
#> 13:  varCatch 2015              var(C) 2.348849e+08 0.90
#> 14:  varCatch 2015              var(C) 2.354604e+08 0.95
#> 15:      varF 2015              var(F) 1.904710e-02 0.05
#> 16:      varF 2015              var(F) 1.927988e-02 0.10
#> 17:      varF 2015              var(F) 1.989139e-02 0.25
#> 18:      varF 2015              var(F) 2.085215e-02 0.50
#> 19:      varF 2015              var(F) 2.130031e-02 0.75
#> 20:      varF 2015              var(F) 2.175045e-02 0.90
#> 21:      varF 2015              var(F) 2.180638e-02 0.95
#>     statistic year                name         data prob

# DEFINE statistics without summaries
statistics <- list(
  CMSY=list(~C/MSY,
    name="CMSY",
    desc="Catch over MSY"))
# COMPUTE performance
perf <- performance(run, statistics, refpts=FLPar(MSY=110000),
  metrics=list(C=catch), years=list(2000:2015))
# COMPUTE summaries
perf[, .(CMSY=mean(data))]

#>        CMSY
#> 1: 1.170685