Compute performance indicators

Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque eleifend odio ac rutrum luctus. Aenean placerat porttitor commodo. Pellentesque eget porta libero. Pellentesque molestie mi sed orci feugiat, non mollis enim tristique. Suspendisse eu sapien vitae arcu lobortis ultrices vitae ac velit. Curabitur id nunc euismod ante fringilla lobortis. Aliquam ullamcorper in diam non placerat.

# S4 method for FLStock
performance(
  x,
  indicators,
  refpts = FLPar(),
  years = as.character(seq(dims(x)$minyear, dims(x)$maxyear)),
  metrics = FLCore::metrics(x),
  probs = NULL,
  mp = NULL
)

# S4 method for FLQuants
performance(
  x,
  indicators,
  refpts = FLPar(),
  years = dims(x[[1]])$maxyear,
  probs = c(0.1, 0.25, 0.5, 0.75, 0.9),
  mp = NULL
)

# S4 method for FLStocks
performance(
  x,
  indicators,
  refpts = FLPar(),
  years = dims(x[[1]])$maxyear,
  metrics = FLCore::metrics,
  probs = NULL,
  grid = missing,
  mp = NULL,
  mc.cores = 1
)

# S4 method for list
performance(
  x,
  indicators,
  refpts = FLPar(),
  years = dims(x[[1]])$maxyear,
  probs = NULL,
  grid = missing,
  mp = NULL,
  mc.cores = 1
)

Arguments

indicators
Indicators to be computed, as formula, name and description, list

refpts
Reference points for calculations, list

years
Years on which indicators should be computed, defaults to last year of input FLQuants

run
Object holding the results of forward projections, as a named FLQuants

Value

data.frame Results of computing performance indicators

Details

Aliquam sagittis feugiat felis eget consequat. Praesent eleifend dolor massa, vitae faucibus justo lacinia a. Cras sed erat et magna pharetra bibendum quis in mi. Sed sodales mollis arcu, sit amet venenatis lorem fringilla vel. Vivamus vitae ipsum sem. Donec malesuada purus at libero bibendum accumsan. Donec ipsum sapien, feugiat blandit arcu in, dapibus dictum felis.

See also

Examples

data(p4om)
indicators <- list(
  T1=list(~yearMeans(C[, -1]/C[, -dims(C)$year]), name="mean(C[t] / C[t-1])",
    desc="Mean absolute proportional change in catch"),
  T2=list(~yearVars(C), name="var(C)", desc="Variance in catch"),
  T3=list(~yearVars(F), name="var(F)", desc="Variance in fishing mortality"))
run <- window(stock(om), start=2000, end=2015)
performance(run, indicators, refpts=FLPar(MSY=0),
  metrics=list(C=catch, F=fbar), years=list(20:25, 20:30))

#>      indicator year data iter                name
#>   1:        T1   25  NaN    1 mean(C[t] / C[t-1])
#>   2:        T1   25  NaN    2 mean(C[t] / C[t-1])
#>   3:        T1   25  NaN    3 mean(C[t] / C[t-1])
#>   4:        T1   25  NaN    4 mean(C[t] / C[t-1])
#>   5:        T1   25  NaN    5 mean(C[t] / C[t-1])
#>  ---                                             
#> 296:        T3   30   NA   46              var(F)
#> 297:        T3   30   NA   47              var(F)
#> 298:        T3   30   NA   48              var(F)
#> 299:        T3   30   NA   49              var(F)
#> 300:        T3   30   NA   50              var(F)