TAC implementation function

Performs a short term forecast (STF) to hit the target F in year ay+1. The resulting catch in year ay+1 is the TAC, i.e. the TAC that will result in Fbar = Ftarget. The STF uses geometric mean recruitment. Fbar in the intermediate years (i.e. years between last data year and ay+1) are set as the mean of the last nsqy years. The control argument is a list of parameters: nsqy - number of years to average over to get Fbar for STF delta_tac_min - constraint on the TAC delta_tac_max - constraint on the TAC

tac.is(
  stk,
  ctrl,
  args,
  dtaclow = NA,
  dtacupp = NA,
  recyrs = 10,
  fmin = 0,
  initac = catch(stk[, ac(iy - 1)]),
  tracking
)

Arguments

stk

The perceived FLStock.

imp_control

A list with the elements: nsqy, delta_tac_min, delta_tac_max

ay

The year for which the target F is set, based on the SSB in year (ay - control$ssb_lag).

EFF0

The tracking array

EFF

Not used by this function but may be used by the other implementation functions

Examples

data(sol274)
#> Warning: namespace ‘DBI’ is not available and has been replaced
#> by .GlobalEnv when processing object ‘om’